### 2020-06-22 #2378 @arfon
Also, while we're waiting for an editor to be assigned, you might want to look at fixing up your citations - the BibTeX key you're using in your `paper.md` seem to be different from the entries in your `paper.bib`.

You can ask Whedon to regenerate the paper proof with the command `@whedon generate pdf`
