+++ paper/paper.bib
@@ -8,7 +8,7 @@
   year={2014}
 }
 
-@inproceedings{10.1145/2889160.2891057,
+@inproceedings{cito2016,
 author = {Cito, J\"{u}rgen and Gall, Harald C.},
 title = {Using Docker Containers to Improve Reproducibility in Software Engineering Research},
 year = {2016},
@@ -25,60 +25,14 @@ location = {Austin, Texas},
 series = {ICSE ’16}
 }
 
-@misc{tensorflow2015-whitepaper,
-title={{TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
-url={http://tensorflow.org/},
-note={Software available from tensorflow.org},
-author={
-    Mart\'{\i}n~Abadi and
-    Ashish~Agarwal and
-    Paul~Barham and
-    Eugene~Brevdo and
-    Zhifeng~Chen and
-    Craig~Citro and
-    Greg~S.~Corrado and
-    Andy~Davis and
-    Jeffrey~Dean and
-    Matthieu~Devin and
-    Sanjay~Ghemawat and
-    Ian~Goodfellow and
-    Andrew~Harp and
-    Geoffrey~Irving and
-    Michael~Isard and
-    Yangqing Jia and
-    Rafal~Jozefowicz and
-    Lukasz~Kaiser and
-    Manjunath~Kudlur and
-    Josh~Levenberg and
-    Dan~Man\'{e} and
-    Rajat~Monga and
-    Sherry~Moore and
-    Derek~Murray and
-    Chris~Olah and
-    Mike~Schuster and
-    Jonathon~Shlens and
-    Benoit~Steiner and
-    Ilya~Sutskever and
-    Kunal~Talwar and
-    Paul~Tucker and
-    Vincent~Vanhoucke and
-    Vijay~Vasudevan and
-    Fernanda~Vi\'{e}gas and
-    Oriol~Vinyals and
-    Pete~Warden and
-    Martin~Wattenberg and
-    Martin~Wicke and
-    Yuan~Yu and
-    Xiaoqiang~Zheng},
-  year={2015},
+@inproceedings{deng2009imagenet,
+  title={Imagenet: A large-scale hierarchical image database},
+  author={Deng, Jia and Dong, Wei and Socher, Richard and Li, Li-Jia and Li, Kai and Fei-Fei, Li},
+  booktitle={2009 IEEE conference on computer vision and pattern recognition},
+  pages={248--255},
+  year={2009},
+  organization={Ieee}
 }
-
-@inproceedings{imagenet_cvpr09,
-        AUTHOR = {Deng, J. and Dong, W. and Socher, R. and Li, L.-J. and Li, K. and Fei-Fei, L.},
-        TITLE = {{ImageNet: A Large-Scale Hierarchical Image Database}},
-        BOOKTITLE = {CVPR09},
-        YEAR = {2009},
-        BIBSOURCE = "http://www.image-net.org/papers/imagenet_cvpr09.bib"}
 
 @article{zaharia2018accelerating,
   title={Accelerating the Machine Learning Lifecycle with MLflow.},
