### 2020-06-22 #2378 @sritchie
Thanks for the note, @arfon! This was a bibtex misunderstanding on my part, as I'm sure you guessed. Let's try this again.
