### 2020-06-22 7240b4b
cloudbuild config and generation script
