### 2020-06-22 7209c90
minor typo corrections
