### 2020-06-23 0b60de4
updates responding to PR review
