+++ scripts/cloudbuild.py
@@ -3,6 +3,7 @@
 
 from absl import flags, app, logging
 from typing import Dict, List, Any, NamedTuple, Optional
+from enum import Enum
 
 import json
 import pprint as pp
@@ -35,15 +36,14 @@ Config = Dict[str, Any]
 # ----------------------------------------------------------------------------
 class BuildStep(NamedTuple):
   '''a google cloud-build step'''
-  sid: str
+  step_id: str
   name: str = _CLOUD_BUILDER
   args: List[str] = []
   dependencies: List[str] = ['-']
 
   def to_dict(self) -> Dict[str, Any]:
-    '''converts step to dict'''
     return {
-        'id': self.sid,
+        'id': self.step_id,
         'name': self.name,
         'args': self.args,
         'waitFor': self.dependencies,
@@ -51,6 +51,12 @@ class BuildStep(NamedTuple):
 
 
 # ----------------------------------------------------------------------------
+class ImageType(Enum):
+  GPU_BASE = 'GPU_BASE'
+  CPU = 'CPU'
+  GPU = 'GPU'
+
+
 class ImageSpec(NamedTuple):
   '''a caliban base image spec'''
   dockerfile: str = _CPU_DOCKERFILE
@@ -59,16 +65,26 @@ class ImageSpec(NamedTuple):
   gpu_version: Optional[str] = None
   build_args: List[str] = []
 
+  @property
+  def image_type(self) -> ImageType:
+    if self.gpu_version is None:
+      return ImageType.CPU
+    else:
+      if self.cpu_version is None:
+        return ImageType.GPU_BASE
+      else:
+        return ImageType.GPU
+
+  @property
   def tag(self) -> str:
-    '''returns tag for this spec'''
     parts = []
-    if self.gpu_version is None:  # cpu image
+    if self.image_type == ImageType.CPU:
       parts = [_CPU_TAG, self.cpu_version]
-    else:  # gpu image
+    else:
       parts.append(_GPU_TAG)
-      if self.cpu_version is None:  # gpu base image
+      if self.image_type == ImageType.GPU_BASE:
         parts += ['base', self.gpu_version]
-      else:  # full gpu image
+      else:
         parts += [self.gpu_version, self.cpu_version]
 
     return f'{FLAGS.base_url}:' + '-'.join(parts)
@@ -76,13 +92,13 @@ class ImageSpec(NamedTuple):
 
 # ----------------------------------------------------------------------------
 def _create_push_step(tag: str) -> BuildStep:
-  '''creates a cloudbuild push step'''
-  return BuildStep(sid=f'{tag}-push', args=['push', tag], dependencies=[tag])
+  return BuildStep(step_id=f'{tag}-push',
+                   args=['push', tag],
+                   dependencies=[tag])
 
 
 # ----------------------------------------------------------------------------
 def _create_gpu_base_image_specs(cfg: Config) -> Dict[str, ImageSpec]:
-  '''creates gpu base image specs'''
   return {
       k: ImageSpec(base_image=None,
                    gpu_version=k,
@@ -93,7 +109,6 @@ def _create_gpu_base_image_specs(cfg: Config) -> Dict[str, ImageSpec]:
 
 # ----------------------------------------------------------------------------
 def _create_cpu_image_specs(cfg: Config) -> Dict[str, ImageSpec]:
-  '''creates cpu image specs'''
   return {k: ImageSpec(cpu_version=k, build_args=v) for k, v in cfg.items()}
 
 
@@ -103,14 +118,13 @@ def _create_image_spec(
     cpu_specs: Dict[str, ImageSpec],
[…]
