### 2020-06-23 4b7b43c
fix default miniconda version in base dockerfile
