### 2020-06-23 7f34b4a
Add cloudbuild configuration for automatic docker image builds (#25)
* cloudbuild config and generation script

* minor typo corrections

* updates responding to PR review

* forgot to add this file to last commit

* fix default miniconda version in base dockerfile

Co-authored-by: Ambrose Slone <aslone@google.com>
