+++ scripts/cloudbuild.py
@@ -1,0 +1,191 @@
+#!/usr/bin/env python
+'''creates cloudbuild.json for caliban'''
+
+from absl import flags, app, logging
+from typing import Dict, List, Any, NamedTuple, Optional
+from enum import Enum
+
+import json
+import pprint as pp
+
+# ----------------------------------------------------------------------------
+_CLOUD_BUILDER = 'gcr.io/cloud-builders/docker'
+_BASE_URL = 'gcr.io/blueshift-playground/blueshift'
+_GPU_BASE_TAG = 'base'
+_CPU_TAG = 'cpu'
+_GPU_TAG = 'gpu'
+_CPU_BASE_IMAGE = 'ubuntu:18.04'
+_CPU_DOCKERFILE = 'dockerfiles/Dockerfile'
+_GPU_DOCKERFILE = 'dockerfiles/Dockerfile.gpu'
+
+# ----------------------------------------------------------------------------
+FLAGS = flags.FLAGS
+
+flags.DEFINE_string('output', None, 'output json file')
+flags.DEFINE_string('config', None, 'input configuration file')
+flags.DEFINE_string('base_url', _BASE_URL, 'base url for images')
+flags.DEFINE_integer('timeout_sec', 24 * 60 * 60,
+                     'timeout for all builds to complete')
+
+# ----------------------------------------------------------------------------
+logging.get_absl_handler().setFormatter(None)
+
+Config = Dict[str, Any]
+
+
+# ----------------------------------------------------------------------------
+class BuildStep(NamedTuple):
+  '''a google cloud-build step'''
+  step_id: str
+  name: str = _CLOUD_BUILDER
+  args: List[str] = []
+  dependencies: List[str] = ['-']
+
+  def to_dict(self) -> Dict[str, Any]:
+    return {
+        'id': self.step_id,
+        'name': self.name,
+        'args': self.args,
+        'waitFor': self.dependencies,
+    }
+
+
+# ----------------------------------------------------------------------------
+class ImageType(Enum):
+  GPU_BASE = 'GPU_BASE'
+  CPU = 'CPU'
+  GPU = 'GPU'
+
+
+class ImageSpec(NamedTuple):
+  '''a caliban base image spec'''
+  dockerfile: str = _CPU_DOCKERFILE
+  base_image: Optional[str] = _CPU_BASE_IMAGE
+  cpu_version: Optional[str] = None
+  gpu_version: Optional[str] = None
+  build_args: List[str] = []
+
+  @property
+  def image_type(self) -> ImageType:
+    if self.gpu_version is None:
+      return ImageType.CPU
+    else:
+      if self.cpu_version is None:
+        return ImageType.GPU_BASE
+      else:
+        return ImageType.GPU
+
+  @property
+  def tag(self) -> str:
+    parts = []
+    if self.image_type == ImageType.CPU:
+      parts = [_CPU_TAG, self.cpu_version]
+    else:
+      parts.append(_GPU_TAG)
+      if self.image_type == ImageType.GPU_BASE:
+        parts += ['base', self.gpu_version]
+      else:
+        parts += [self.gpu_version, self.cpu_version]
+
+    return f'{FLAGS.base_url}:' + '-'.join(parts)
+
+
+# ----------------------------------------------------------------------------
+def _create_push_step(tag: str) -> BuildStep:
+  return BuildStep(step_id=f'{tag}-push',
+                   args=['push', tag],
+                   dependencies=[tag])
+
+
+# ----------------------------------------------------------------------------
+def _create_gpu_base_image_specs(cfg: Config) -> Dict[str, ImageSpec]:
+  return {
+      k: ImageSpec(base_image=None,
+                   gpu_version=k,
[…]
