### 2020-06-23 e10795a
documentation updates for base images
