### 2020-06-23 dfb26c5
fix pre-commit trailing whitespace error
