### 2020-06-24 4653ba1
added link to gcr base image repo to docs
