### 2020-06-24 4dc922a
adds strict_db flag to run/cloud/cluster commands
