### 2020-06-24 9eb5913
added some detail about base image tag formatting
