### 2020-06-24 74788a1
Aslone/cloudbuild docs (#27)
* documentation updates for base images

* fix pre-commit trailing whitespace error

* added link to gcr base image repo to docs

* added some detail about base image tag formatting

* fixed typo
