+++ docs/explore/base_image.rst
@@ -10,23 +10,88 @@ environment inside of each container. This guarantees you a truly blank slate;
 the dependencies you declare in your code directory are the only Python
 libraries that will be present. No more version clashes or surprises.
 
-Caliban uses one of two base images, depending on whether you're running in GPU
-(default) or CPU mode:
+Caliban uses a set of base images covering a set of common combinations of
+python and cuda versions. You can find our base images
+`here <https://pantheon.corp.google.com/gcr/images/blueshift-playground/GLOBAL/blueshift>`_.
+The format of our base image names is ``gcr.io/blueshift-playground/blueshift:TAG``,
+where ``TAG`` describes the configuration of the base image.
 
-* ``gcr.io/blueshift-playground/blueshift:gpu`` for the default GPU mode
-* ``gcr.io/blueshift-playground/blueshift:cpu`` for CPU, or ``--nogpu``\ , mode
+For example, ``gcr.io/blueshift-playground/blueshift:gpu-cuda100-py37`` is a
+base image that has CUDA 10.0 and python 3.7, while
+``gcr.io/blueshift-playground/blueshift:cpu-py38`` is a base image that has
+no CUDA support and uses python 3.8.
 
-These are based on, respectively,
+Our current matrix of supported combinations:
 
-* ``tensorflow/tensorflow:2.2.0-gpu-py3``
-* ``tensorflow/tensorflow:2.2.0-py3``
++-----------+------------+------------+
+|           | python 3.7 | python 3.8 |
++-----------+------------+------------+
+| no cuda   |    yes     |    yes     |
++-----------+------------+------------+
+| cuda 10.0 |    yes     |    yes     |
++-----------+------------+------------+
+| cuda 10.1 |    yes     |    yes     |
++-----------+------------+------------+
 
-We chose the base Tensorflow containers only because they do the hard work of
-installing all of the CUDA drivers and other software required by NVIDIA GPUs;
-the virtual environment inside of the container isolates you from the installed
-``tensorflow`` library. You can install any TF version you like, or use Jax or
-Pytorch or any other system.
+These images are automatically updated, and if you have an image combination that
+we don't support, please file an issue and we'll consider adding it to our set
+of supported images. We are planning to add support for custom base images so
+you can build and use your own specialized image.
 
-Here's a link to the `Dockerfile
-<https://github.com/google/caliban/blob/master/Dockerfile>`_ we use to build the
-two base images that sit behind all Docker images generated by Caliban.
+The dockerfiles we use to generate our supported images can be found
+`here <https://github.com/google/caliban/tree/master/dockerfiles>`_. We create
+base gpu images from the `Dockerfile.gpu <https://github.com/google/caliban/blob/master/dockerfiles/Dockerfile.gpu>`_
+file, and then use these as base images for creating full GPU images with
+support for specific python versions using this `Dockerfile <https://github.com/google/caliban/blob/master/dockerfiles/Dockerfile>`_.
+
+We base our gpu base images on the `nvidia/cuda <https://hub.docker.com/r/nvidia/cuda/>`_
+images, which contain the relevant CUDA drivers required for GPU use. The virtual
+environment inside of the Caliban container isolates you from these low-level details,
+so you can install any tensorflow version you like, or use Jax or Pytorch or any
+other system.
+
+Details for Maintainers
+~~~~~~~~~~~~~~~~~~~~~~~
+
+We utilize Google's `Cloud Build <http://cloud.google.com/cloud-build/docs>`_ service
+to build Caliban's base images. Our Cloud Build configuration file that controls
+our image generation can be found in the source repository
+`here <https://github.com/google/caliban/blob/master/cloudbuild.json>`_.
+
+This file can quickly get lengthy and difficult to maintain, so we generate this file
[…]
