### 2020-06-25 ce0eca9
Flush buffered docker build stdout + fix close error from #29 (#30)
