### 2020-06-25 #2378 @diehlpk
> @diehlpk - no one will ever complain if you volunteer to edit :)
>
> And you can just claim the paper by issuing the same command I did, or even `@whedon assign me as editor`

Ok, will do. I just did not want to overpass the Managing EiC.
