### 2020-06-25 #2378 @diehlpk
Hi @terrytangyuan  or @arokem or @kmoham6 or @stevenrbrandt or @rtohid  would one of you be interested to review this paper?
