### 2020-06-26 #2378 @diehlpk
@cboettig Would you be interested to review this paper?
