### 2020-06-26 179d71f
get progress bar working
