+++ caliban/util.py
@@ -38,6 +38,7 @@ from typing import (Any, Callable, Dict, Iterable, List, NamedTuple, Optional,
 import tqdm
 from absl import logging
 from blessings import Terminal
+from tqdm._utils import _term_move_up
 
 t = Terminal()
 
@@ -402,22 +403,34 @@ def capture_stdout(cmd: List[str],
 
   buf = io.StringIO()
   ret_code = None
+  prefix = _term_move_up() + '\r'
+
   with subprocess.Popen(cmd,
                         stdin=subprocess.PIPE,
                         stdout=subprocess.PIPE,
                         stderr=subprocess.STDOUT,
-                        bufsize=1,
-                        encoding="utf-8") as p:
+                        universal_newlines=False,
+                        bufsize=1) as p:
     if input_str:
-      p.stdin.write(input_str)
+      p.stdin.write(input_str.encode('utf-8'))
     p.stdin.close()
 
-    for line in p.stdout:
-      print(line, end='', file=file)
+    out = io.TextIOWrapper(p.stdout, newline='')
+    carriage_pending = False
+
+    for line in out:
       buf.write(line)
 
-    # flush to force the contents to display.
-    file.flush()
+      if carriage_pending:
+        line = prefix + line
+        carriage_pending = False
+
+      if line.endswith('\r'):
+        carriage_pending = True
+        line = line[:-1] + '\n'
+
+      file.write(line)
+      file.flush()
 
     while p.poll() is None:
       # Process hasn't exited yet, let's wait some
@@ -676,8 +689,7 @@ class TqdmFile(object):
     self.file = file
 
   def write(self, x):
-    if len(x.rstrip()) > 0:
-      tqdm.tqdm.write(x, file=self.file, end='')
+    tqdm.tqdm.write(x, file=self.file, end='')
 
   def flush(self):
     return getattr(self.file, "flush", lambda: None)()
