### 2020-06-27 #2378 @TheChymera
@diehlpk thank you for thinking of me :) While I am indeed interested in new work on software reproducibility, it is my constant experience that Docker is an incredibly unfitting and shoehorned solution for this. It's a deployment system, and has nothing to do with environment reproducibility as much as copy-paste-ability. I am afraid I could not offer a competent review beyond just disagreeing very strongly with the fundamental premise and the ex ante choice of technologies.
