### 2020-06-27 #2378 @sritchie
@TheChymera I actually agree with you, as far as raw Docker is concerned; it may not be clear, but I built this as a reaction to the dozens of copy-pasted Dockerfiles that end up infesting projects that attempt to use this tool to give themselves reproducibility guarantees.

Caliban is an attempt to implement a declarative build system that compiles a user's project on top of a stable base image. It uses Docker to manage caching and the stable base image, but relies more on tools like conda and virtualenv to provide reproducibility guarantees.

I didn't see a way (or a need) to get away from Docker, because Docker really is the only way to ship an environment from a local machine to a Cloud environment. All of the Cloud providers either have you ship a Docker image, or they attempt to reproduce the environment for you, which inevitably breaks if you're developing on a Mac and shipping to Linux.

So, instead of going down that bad road, we built a tool that piggybacks on tools you already have to use — some declaration of requirements, and a directory of code — to dynamically generate the usually-copy-pasted code. The fact that this lets you locally execute the exact same image and environment that the cloud provider will execute is really valuable, imo.

Maybe you're the perfect reviewer! Especially if you have some alternative road to the goal, here.
