### 2020-06-29 #2403 @whedon
[REVIEW]: Caliban: Docker-based job manager for reproducible workflows
**Submitting author:** @sritchie (<a href="http://orcid.org/0000-0002-0545-6360">Sam Ritchie</a>)
**Repository:** <a href="https://github.com/google/caliban" target ="_blank">https://github.com/google/caliban</a>
**Version:** 0.4.1
