### 2020-06-30 #2403 @lukasheinrich
hi @sritchie -- part of the described functionality requires Google Cloud Access. Will you provide credits or similar to allow us to test that functionality (I don't have any free credits left)?
