+++ docs/gke/prereq.rst
@@ -8,5 +8,5 @@ Required Permissions
 
 To create and use a GKE cluster, you'll need to modify your service account key
 to give it Account Owner permissions. Those instructions live at the
-:ref:`Service Account Key` docs page. Note that this only applies if you are
+:doc:`/cloud/service_account` docs page. Note that this only applies if you are
 using a service account key.
