### 2020-07-09 85c0bc3
Add support for gcloud auth login to enable job submission credentials (#34)
