+++ tutorials/uv-metrics/mnist.py
@@ -1,0 +1,77 @@
+#!/usr/bin/python
+#
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""This tutorial comes from the Tensorflow MNIST quickstart at
+https://www.tensorflow.org/tutorials/quickstart/beginner.
+
+"""
+import warnings
+
+import tensorflow as tf
+from absl import app
+
+import cli
+
+warnings.filterwarnings("ignore", category=DeprecationWarning)
+
+
+def get_keras_model(width=128, activation='relu'):
+  """Returns an instance of a Keras Sequential model.
+https://www.tensorflow.org/api_docs/python/tf/keras/Sequential"""
+  return tf.keras.models.Sequential([
+      tf.keras.layers.Flatten(input_shape=(28, 28)),
+      tf.keras.layers.Dense(width, activation=activation),
+      tf.keras.layers.Dense(width, activation=activation),
+      tf.keras.layers.Dense(10, activation=None),
+  ])
+
+
+def model_main(learning_rate=0.01, epochs=3, **kwargs):
+  """Train a model against the MNIST dataset and print performance metrics."""
+  mnist = tf.keras.datasets.mnist
+
+  (x_train, y_train), (x_test, y_test) = mnist.load_data()
+  x_train, x_test = x_train / 255.0, x_test / 255.0
+
+  model = get_keras_model()
+
+  loss_fn = tf.keras.losses.SparseCategoricalCrossentropy(from_logits=True)
+  optimizer = tf.keras.optimizers.Adam(learning_rate=learning_rate)
+
+  model.compile(optimizer=optimizer, loss=loss_fn, metrics=['accuracy'])
+
+  print(
+      f'Training model with learning rate={learning_rate} for {epochs} epochs.')
+  model.fit(x_train, y_train, epochs=epochs)
+
+  print('Model performance: ')
+  model.evaluate(x_test, y_test)
+
+
+def run_app(args):
+  """Main function to begin training."""
+  model_main(learning_rate=args.learning_rate,
+             epochs=args.epochs,
+             local_path=args.local_path,
+             gcloud_path=args.gcloud_path,
+             tensorboard_path=args.tensorboard_path)
+
+
+def main():
+  app.run(run_app, flags_parser=cli.parse_flags)
+
+
+if __name__ == '__main__':
+  main()
