+++ tutorials/uv-metrics/trainer/train.py
@@ -1,0 +1,459 @@
+"""This module, and plot.py in the same folder, demonstrate how to use UV's
+reporter and reader classes to collect data during a training run and then load
+it again after the run for analysis.
+
+"""
+
+from __future__ import division, print_function
+
+import getpass
+import os
+from typing import Dict, Optional
+
+import fs
+import mlflow
+import tensorflow as tf
+import tensorflow_datasets as tfds
+import tqdm
+import uv.reporter as r
+import uv.tensorflow.reporter as rft
+import uv.util as u
+from absl import app
+from uv.fs.reporter import FSReporter
+from uv.mlflow.reporter import MLFlowReporter
+
+from trainer import cli
+
+MNIST = "mnist"
+
+
+def normalize(example):
+  """This normalizes all color values to be between 0 and 1. This gets turned
+  into a tensorflow operation... print statements will only get called a single
+  time.
+
+  """
+  example['image'] = tf.cast(example['image'], tf.float32) / 255.
+  return example
+
+
+def prepare_mnist(
+    batch_size: int,
+    train_shuffle_buffer: int = 1000) -> Dict[str, tf.data.Dataset]:
+  """Generates a dictionary with keys measure, train, test, all from mnist.
+
+  https://www.tensorflow.org/datasets/catalog/mnist
+
+  Training has 60k examples, while test has 10k examples.
+
+  """
+  train = tfds.load(MNIST, split=tfds.Split.TRAIN,
+                    shuffle_files=True).map(normalize)
+  test = tfds.load(MNIST, split=tfds.Split.TEST).map(normalize)
+
+  return {
+      # This will create a buffer of `train_shuffle_buffer` items in memory,
+      # then sample from these to get samples. This biases toward the first
+      # batch of examples. the "repeat" effectively lets you go forever.
+      #
+      # If you take enough batches to run out the epoch, you'll start pulling
+      # from the beginning again.
+      "batched": train.shuffle(train_shuffle_buffer).repeat().batch(batch_size),
+
+      # These are the unbatched train and test sets, used for evaluation and
+      # metric generation at the end of each batch.
+      "train": train,
+      "test": test
+  }
+
+
+def build_model(activation: str, width: int, depth: int) -> tf.keras.Model:
+  """Return the bare model that we'll train."""
+  out = inp = tf.keras.layers.Input(shape=(28, 28, 1))
+  out = tf.keras.layers.Flatten()(out)
+
+  # Stack a bunch of layers of the same width.
+  for _ in range(depth):
+    out = tf.keras.layers.Dense(width, activation=activation)(out)
+
+  # Append a final layer that checks 10 classes.
+  out = tf.keras.layers.Dense(10, activation=None)(out)
+
+  # return the final model, all sewn together.
+  return tf.keras.Model(inputs=inp, outputs=out)
+
+
+@tf.function
+def compute_loss(labels, logits):
+  """Compute loss for a single batch of data, given the precomputed logits and
+  expected labels. The returned loss is normalized by the batch size.
+
+  """
+  cur_batch_size = tf.cast(labels.shape[0], tf.float32)
+  cce = tf.losses.SparseCategoricalCrossentropy(
+      from_logits=True, reduction=tf.losses.Reduction.SUM)
+
+  # loss here is the cce, normalized by the batch size.
+  return cce(labels, logits) / cur_batch_size
+
+
+@tf.function
+def compute_accuracy(labels, logits):
+  """Compute accuracy for a single batch of data, given the precomputed logits
+  and expected labels. The returned accuracy is normalized by the batch size.
+
+  """
+  current_batch_size = tf.cast(labels.shape[0], tf.float32)
+
+  # logits is the percent chance; this gives the category for each.
+  predictions = tf.argmax(logits, axis=1)
+
+  # return the average number of items equal to their label.
[…]
