+++ tests/caliban/history/test_history.py
@@ -13,36 +13,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-#!/usr/bin/python
-#
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 """unit tests for caliban history"""
 
-import pytest  #type: ignore
-# https://mypy.readthedocs.io/en/latest/jobning_mypy.html#missing-imports
+from datetime import datetime
 
 from sqlalchemy.engine.base import Engine
-from caliban.history.utils import session_scope, get_mem_engine, get_sql_engine
-from caliban.history.types import (JobStatus, Platform, ExperimentGroup,
-                                   Experiment, JobSpec, Job, ContainerSpec)
 
+import pytest  # type: ignore
+from caliban.history.types import (ContainerSpec, Experiment, ExperimentGroup,
+                                   Job, JobSpec, Platform)
+from caliban.history.util import get_mem_engine, session_scope
 from caliban.util import current_user
-import random
-from datetime import datetime
-from time import sleep
+
+# https://mypy.readthedocs.io/en/latest/jobning_mypy.html#missing-imports
 
 # we create and exist session scopes here to test persistence
 
