+++ caliban/builder.py
@@ -1,729 +1,0 @@
-#!/usr/bin/python
-#
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Functions required to interact with Docker to build and run images, shells
-and notebooks in a Docker environment.
-
-"""
-
-from __future__ import absolute_import, division, print_function
-
-import json
-import os
-import subprocess
-from enum import Enum
-from pathlib import Path
-from typing import (Any, Callable, Dict, Iterable, List, NamedTuple, NewType,
-                    Optional, Union)
-
-from absl import logging
-from blessings import Terminal
-
-import caliban.config as c
-import caliban.util as u
-import caliban.util.fs as ufs
-from caliban.history.types import JobSpec
-
-t = Terminal()
-
-DEV_CONTAINER_ROOT = "gcr.io/blueshift-playground/blueshift"
-TF_VERSIONS = {"2.2.0", "1.12.3", "1.14.0", "1.15.0"}
-DEFAULT_WORKDIR = "/usr/app"
-CREDS_DIR = "/.creds"
-CONDA_BIN = "/opt/conda/bin/conda"
-
-ImageId = NewType('ImageId', str)
-ArgSeq = NewType('ArgSeq', List[str])
-
-
-class DockerError(Exception):
-  """Exception that passes info on a failed Docker command."""
-
-  def __init__(self, message, cmd, ret_code):
-    super().__init__(message)
-    self.message = message
-    self.cmd = cmd
-    self.ret_code = ret_code
-
-  @property
-  def command(self):
-    return " ".join(self.cmd)
-
-
-class NotebookInstall(Enum):
-  """Flag to decide what to do ."""
-  none = 'none'
-  lab = 'lab'
-  jupyter = 'jupyter'
-
-  def __str__(self) -> str:
-    return self.value
-
-
-class Shell(Enum):
-  """Add new shells here and below, in SHELL_DICT."""
-  bash = 'bash'
-  zsh = 'zsh'
-
-  def __str__(self) -> str:
-    return self.value
-
-
-# Tuple to track the information required to install and execute some custom
-# shell into a container.
-ShellData = NamedTuple("ShellData", [("executable", str),
-                                     ("packages", List[str])])
-
-
-def apt_install(*packages: str) -> str:
-  """Returns a command that will install the supplied list of packages without
-  requiring confirmation or any user interaction.
-  """
-  package_str = ' '.join(packages)
-  no_prompt = "DEBIAN_FRONTEND=noninteractive"
-  return f"{no_prompt} apt-get install --yes --no-install-recommends {package_str}"
-
-
-def apt_command(commands: List[str]) -> List[str]:
-  """Pre-and-ap-pends the supplied commands with the appropriate in-container and
-  cleanup command for aptitude.
-
-  """
-  update = ["apt-get update"]
-  cleanup = ["apt-get clean", "rm -rf /var/lib/apt/lists/*"]
-  return update + commands + cleanup
-
-
-# Dict linking a particular supported shell to the data required to run and
-# install the shell inside a container.
-#
-# : Dict[Shell, ShellData]
-SHELL_DICT = {
[…]
