+++ caliban/platform/gke/cli.py
@@ -31,10 +31,10 @@ import caliban.config as conf
 import caliban.platform.gke.constants as k
 import caliban.platform.gke.util as util
 import caliban.util as u
-from caliban.platform.cloud.core import generate_image_tag
-from caliban.platform.gke.cluster import Cluster
 from caliban.history.util import (create_experiments, generate_container_spec,
                                   get_mem_engine, get_sql_engine, session_scope)
+from caliban.platform.cloud.core import generate_image_tag
+from caliban.platform.gke.cluster import Cluster
 
 
 # ----------------------------------------------------------------------------
