+++ tests/caliban/docker/test_push.py
@@ -27,3 +27,53 @@ def test_image_tag_for_project():
 
   assert p._image_tag_for_project(
       "google.com:face", "imageid") == "gcr.io/google.com/face/imageid:latest"
+
+
+def test_force_push_uuid_tag(fake_process):
+  """Check that the push command actually attempts to tag and push."""
+  project_id = "project"
+  image_id = "imageid"
+
+  tag = p._image_tag_for_project(project_id, image_id)
+
+  fake_process.register_subprocess(["docker", "tag", image_id, tag])
+  fake_process.register_subprocess(["docker", "push", tag])
+
+  assert p.push_uuid_tag(project_id, image_id, force=True) == tag
+
+
+def test_already_pushed_uuid_tag(fake_process):
+  """Check that push_uuid_tag does NOT attempt to push if the process already
+  exists.."""
+  project_id = "project"
+  image_id = "imageid"
+
+  base_tag = p._image_tag_for_project(project_id, image_id, include_tag=False)
+  tag = p._image_tag_for_project(project_id, image_id)
+
+  fake_process.register_subprocess(
+      ["gcloud", "container", "images", "list-tags", "--format=json", base_tag],
+      stdout="[{\"metadata\": []}]")
+
+  assert p.push_uuid_tag(project_id, image_id) == tag
+
+
+def test_push_uuid_tag_if_no_remote_image(fake_process):
+  """Check that push_uuid_tag DOES attempt to push if the image doesn't exist in
+  the remote container registry already.
+
+  """
+  project_id = "project"
+  image_id = "imageid"
+
+  base_tag = p._image_tag_for_project(project_id, image_id, include_tag=False)
+  tag = p._image_tag_for_project(project_id, image_id)
+
+  fake_process.register_subprocess(
+      ["gcloud", "container", "images", "list-tags", "--format=json", base_tag],
+      stdout="[]")
+
+  fake_process.register_subprocess(["docker", "tag", image_id, tag])
+  fake_process.register_subprocess(["docker", "push", tag])
+
+  assert p.push_uuid_tag(project_id, image_id) == tag
