### 2020-07-13 36ff6b1
add basic calibanconfig schema
