### 2020-07-13 6edb9b5
Project restructure to make it easier to implement more backends (#32)
