+++ tests/caliban/docker/test_push.py
@@ -13,3 +13,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+import caliban.docker.push as p
+
+
+def test_image_tag_for_project():
+  """Tests that we generate a valid image tag for domain-scoped and modern
+    project IDs.
+
+    """
+  assert p._image_tag_for_project("face",
+                                  "imageid") == "gcr.io/face/imageid:latest"
+
+  assert p._image_tag_for_project(
+      "google.com:face", "imageid") == "gcr.io/google.com/face/imageid:latest"
