### 2020-07-13 #2403 @diehlpk
Hey @sritchie I think you are fine and the reviewers will proceed with their review. Normally, the review time is about 6 weeks, since we asked the reviewers to finish their review within 6 weeks.
