### 2020-07-15 c1f9aa8
convert argparse to schema
