### 2020-07-15 baa3eeb
remove unused functions
