+++ caliban/config/__init__.py
@@ -17,15 +17,12 @@
 Utilities for our job runner, for working with configs.
 """
 
-import argparse
 import os
 import sys
 from enum import Enum
 from typing import Any, Dict, List, Optional
 
-import commentjson
 import schema as s
-import yaml
 
 import caliban.platform.cloud.types as ct
 import caliban.util.schema as us
