### 2020-07-15 850a049
Add Schema for Calibanconfig (#37)
