### 2020-07-15 fe9d857
fix for issue #40
