### 2020-07-16 96cc987
add custom base image support
