+++ tests/caliban/config/test_config.py
@@ -1,4 +1,5 @@
 #!/usr/bin/python
+
 #
 # Copyright 2020 Google LLC
 #
@@ -18,16 +19,82 @@ import json
 import os
 from argparse import ArgumentTypeError
 
+import hypothesis.strategies as st
+from hypothesis import given
+
 import caliban.config as c
 import caliban.platform.cloud.types as ct
 import caliban.util.schema as us
 import pytest
 
 
+def test_parse_job_mode():
+  assert c.JobMode.parse('CpU') == c.JobMode.CPU
+  assert c.JobMode.parse('cpu') == c.JobMode.CPU
+  assert c.JobMode.parse('cpU  ') == c.JobMode.CPU
+
+  assert c.JobMode.parse('  GpU') == c.JobMode.GPU
+  assert c.JobMode.parse('gpu') == c.JobMode.GPU
+  assert c.JobMode.parse('GPU') == c.JobMode.GPU
+
+  with pytest.raises(Exception):
+    c.JobMode.parse('random')
+
+
+@given(st.text() | st.sampled_from(sorted(c.DLVM_CONFIG.keys())))
+def test_expand_image(s):
+  """Expanded images either exist in the DLVM_CONFIG, or are round-tripped
+  without getting changed.
+
+  """
+  if s in c.DLVM_CONFIG:
+    assert c.expand_image(s) == c.DLVM_CONFIG[s]
+  else:
+    assert c.expand_image(s) == s
+
+
 def test_gpu():
   assert c.gpu(c.JobMode.GPU)
   assert not c.gpu(c.JobMode.CPU)
   assert not c.gpu("face")
+
+
+def test_extract_script_args():
+  # Basic cases. If there are NO script args, or if the default is present,
+  # they're passed back out.
+  assert c.extract_script_args({}) is None
+  assert c.extract_script_args({'script_args': None}) is None
+  assert c.extract_script_args({'script_args': []}) == []
+
+  args = ["--carrot", "stick"]
+
+  # If a '--' is passed in at the head it's stripped off and ignored.
+  assert c.extract_script_args({"script_args": ["--"] + args}) == args
+  assert c.extract_script_args({"script_args": args}) == args
+
+
+def test_extract_project_id(monkeypatch):
+  if os.environ.get('PROJECT_ID'):
+    monkeypatch.delenv('PROJECT_ID')
+
+  # if NO project ID is specified on the environment OR in the supplied config,
+  # the system attempts to exit.
+  with pytest.raises(SystemExit) as wrapped_e:
+    c.extract_project_id({})
+
+  assert wrapped_e.type == SystemExit
+  assert wrapped_e.value.code == 1
+
+  # the project ID gets mirrored back if it exists in the config.
+  assert c.extract_project_id({'project_id': "face"}) == "face"
+
+  monkeypatch.setenv('PROJECT_ID', "env_id")
+
+  # If the env variable is set it's returned:
+  assert c.extract_project_id({'project_id': None}) == "env_id"
+
+  # Unless the project ID's set in the config.
+  assert c.extract_project_id({'project_id': "face"}) == "face"
 
 
 def test_extract_region(monkeypatch):
@@ -52,16 +119,64 @@ def test_extract_region(monkeypatch):
   assert c.extract_region({"region": "us-west1"}) == ct.US.west1
 
 
+def test_extract_cloud_key(monkeypatch):
+  k = 'GOOGLE_APPLICATION_CREDENTIALS'
+  if os.environ.get(k):
+    monkeypatch.delenv(k)
+
+  # initial missing case.
+  assert c.extract_cloud_key({}) == None
+
[…]
