### 2020-07-16 f95b955
Custom base image support in calibanconfig (#39)
