+++ caliban/config/__init__.py
@@ -20,7 +20,7 @@ Utilities for our job runner, for working with configs.
 import os
 import sys
 from enum import Enum
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Tuple
 
 import schema as s
 
@@ -35,7 +35,7 @@ class JobMode(str, Enum):
 
   @staticmethod
   def parse(label):
-    return JobMode(label.upper())
+    return JobMode(label.upper().strip())
 
 
 DRY_RUN_FLAG = "--dry_run"
@@ -58,7 +58,53 @@ DEFAULT_ACCELERATOR_CONFIG = {
     "type": "ACCELERATOR_TYPE_UNSPECIFIED"
 }
 
+# Dictionary of the DLVM "Platform" to a sequence of versions that are
+# currently available as DLVMs. The full list of images is here:
+# https://console.cloud.google.com/gcr/images/deeplearning-platform-release/GLOBAL/
+DLVMS = {
+    "pytorch": [None, "1.0", "1.1", "1.2", "1.3", "1.4"],
+    "tf": [None, "1.0", "1.13", "1.14", "1.15"],
+    "tf2": [None, "2.0", "2.1", "2.2"],
+}
+
 # Schema for Caliban Config
+
+
+def _dlvm_config(job_mode: JobMode) -> Dict[str, str]:
+  """Generates a dict of custom DLVM image identifier -> the actual image ID
+  available from GCR.
+
+  """
+  mode = job_mode.lower()
+
+  def with_version(s: str, version: Optional[str], sep: str) -> Tuple[str, str]:
+    return f"{s}{sep}{version}" if version else s
+
+  def image(lib: str, version: Optional[str]) -> str:
+    base = f"gcr.io/deeplearning-platform-release/{lib}-{mode}"
+    k = with_version(f"dlvm:{lib}-{mode}", version, "-")
+    v = with_version(base, version.replace('.', '-') if version else None, ".")
+    return (k, v)
+
+  return dict(
+      [image(lib, v) for lib, versions in DLVMS.items() for v in versions])
+
+
+# This is a dictionary of some identifier like 'dlvm:pytorch-1.0' to the actual
+# Docker image ID.
+DLVM_CONFIG = {
+    **_dlvm_config(JobMode.CPU),
+    **_dlvm_config(JobMode.GPU),
+}
+
+
+def expand_image(image: str) -> str:
+  """If the supplied image is one of our special prefixed identifiers, returns
+  the expanded Docker image ID. Else, returns the input.
+
+  """
+  return DLVM_CONFIG.get(image, image)
+
 
 AptPackages = s.Or(
     [str], {
@@ -67,24 +113,24 @@ AptPackages = s.Or(
     },
     error=""""apt_packages" entry must be a dictionary or list, not '{}'""")
 
-CalibanConfig = s.Schema({
-    s.Optional("build_time_credentials", default=False):
-        bool,
-    s.Optional("default_mode", default=JobMode.CPU):
-        s.Use(JobMode.parse),
-    s.Optional("project_id"):
-        s.And(str, len),
-    s.Optional("cloud_key"):
-        s.And(str, len),
-    s.Optional("base_image"):
-        str,
-    s.Optional("apt_packages", default=dict):
-        AptPackages,
+Image = s.And(str, s.Use(expand_image))
 
-    # Allow extra entries without killing the schema to allow for backwards
-    # compatibility.
-    s.Optional(str):
-        str,
+BaseImage = s.Or(
+    Image, {
+        s.Optional("gpu", default=None): Image,
+        s.Optional("cpu", default=None): Image
+    },
+    error=
[…]
