### 2020-07-17 a091696
Thread Julia options through configuration file
