### 2020-07-17 6dbfc1c
Begin to work on Slurm backend
