### 2020-07-18 75480d8
Slurm: Allow specifying a job name
