### 2020-07-20 57be6c3
Finish prototype implementation
