+++ caliban/platform/slurm/cli.py
@@ -22,28 +22,40 @@ import pprint as pp
 import re
 import shlex
 import subprocess
+from blessings import Terminal
 from datetime import datetime
+from pprint import pformat
 from typing import Any, Dict, List, Optional
 
 import caliban.cli as cli
 import caliban.config as conf
+import caliban.docker.build as db
 import caliban.util as u
 from caliban.history.util import (create_experiments, generate_container_spec,
                                   get_mem_engine, get_sql_engine, session_scope)
+from caliban.platform.cloud.core import generate_image_tag
+
+
+t = Terminal()
 
 
 # ----------------------------------------------------------------------------
 # Remote access via ssh
+_ssh = "ssh"
 _hostname = "symmetry.pi.local"
 _username = "eschnetter"
 
+# Environment modules
+_setup_cmds = [["source", "/etc/profile"],
+               ["module", "load", "slurm"]]
+
 # Job management via Slurm
+_sbatch = "sbatch"
+_sinfo = "sinfo"
+_squeue = "squeue"
 _partition = "debugq"
 _timelimit = "00:10:00"
-_slurm_path = "/cm/shared/apps/slurm/19.05.5"
-_sbatch = _slurm_path + "/bin/sbatch"
-_sinfo = _slurm_path + "/bin/sinfo"
-_squeue = _slurm_path + "/bin/squeue"
+_nodes = 1
 
 
 # ----------------------------------------------------------------------------
@@ -103,9 +115,29 @@ def _job_submit(args: dict) -> None:
   args: argument dictionary
   """
 
+  script_args = conf.extract_script_args(args)
+  job_mode = cli.resolve_job_mode(args)
+  docker_args = cli.generate_docker_args(job_mode, args)
+  docker_run_args = args.get('docker_run_args', []) or []
+  dry_run = args['dry_run']
+  package = args['module']
+  job_name = args.get('name')
+  gpu_spec = args.get('gpu_spec')
+  preemptible = not args['nonpreemptible']
+  min_cpu = args.get('min_cpu')
+  min_mem = args.get('min_mem')
+  experiment_config = args.get('experiment_config') or [{}]
+  xgroup = args.get('xgroup')
+  image_tag = args.get('image_tag')
+  export = args.get('export', None)
+
+  labels = args.get('label')
+  if labels is not None:
+    labels = dict(cu.sanitize_labels(args.get('label')))
+
+  # Generate job name
   dt = datetime.now().astimezone()
   job_prefix = f'caliban-{dt.strftime("%Y%m%d-%H%M%S")}'
-  job_name = args.get('name')
   # TODO: Sanitize job_name
   if job_name is None:
     job_name = job_prefix
@@ -113,42 +145,104 @@ def _job_submit(args: dict) -> None:
   else:
     output_filename = job_prefix + "-" + job_name + ".log"
 
-  script = """\
-#!/bin/bash
-# echo 'Hello, World!'
-# echo Date: $(date)
-# echo Hostname: $(hostname)
-/cm/shared/apps/singularity/3.3.0/bin/singularity run docker://godlovedc/lolcow
-"""
+  # Arguments to internally build the image required to submit to Slurm
+  docker_m = {'job_mode': job_mode, 'package': package, **docker_args}
 
-  logging.info("Submitting job...")
-  cmd = [_sbatch,
-         "--job-name", job_name,
-         "--nodes", "1",
-         "--output", output_filename,
-         "--partition", _partition,
-         "--time", _timelimit]
-  process = subprocess.run(_with_ssh(cmd),
-                           input=script,
-                           stdout=subprocess.PIPE,
-                           universal_newlines=True)
-  output = process.stdout.rstrip()
[…]
