### 2020-07-20 fbdcd03
Reformat code
