### 2020-07-22 e765d7f
Slurm: Make proper job submission work
