### 2020-07-22 google/caliban#48 @arokem
[JOSS review] community guidelines
For this check-list item:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No. 1 is already in place, but I don't see clear guidelines for no. 2 and 3.
