### 2020-07-22 #2403 @arokem
@diehlpk : I am making slow progress through the checklist. My main remaining items are about testing the various functionality of the library. So far, it's been an enjoyable experience 😄

Regarding community guidelines, I've posted an issue [in the software repo](https://github.com/google/caliban/issues/48).

I also have a small comment for this item:

> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The authors tend to make liberal use of superlatives (the word "tremendous" appears more than once). I think that it would be good to use these sparingly.
