### 2020-07-22 #2403 @arokem
One more comment, regarding the state of the art. There are two strands of research not currently cited, that I believe could be connected to this software and would better embed the software in context:

The first regards automated docker deployment for scientific software. I am thinking primarily of [repo2docker](https://github.com/jupyter/repo2docker), which implements functionality similar to `caliban notebook` and [binder](https://mybinder.org). This paper could be cited: https://openreview.net/forum?id=BJlR6KTE3X

The second regards automated deployment of scientific computing to the cloud. One set of work there is [pywren](http://pywren.io/) and related papers (e.g., https://shivaram.org/publications/pywren-socc17.pdf).  The other is our own work on [cloudknot](https://nrdg.github.io/cloudknot/). If you feel so compelled, you could cite our scipy paper on that: http://conference.scipy.org/proceedings/scipy2018/adam_richie-halford.html.
