### 2020-07-22 google/caliban#49 @arokem
Feature request: support REES
REES is the specification supported by [repo2docker](https://repo2docker.readthedocs.io/en/latest/specification.html). It would be great to support that. In particular, as I understand it (but correct me if I am wrong), caliban currently supports dependencies specified in a `requirements.txt` file. REES expands this to support dependencies in a number of other formats, including conda `environment.yml` files and even `Dockerfile`s. Implementing this would increase the range of things users could do with the software.
