### 2020-07-22 #2403 @arokem
Relatedly, but definitely not required for the review, some food for thought here: https://github.com/google/caliban/issues/49
