### 2020-07-22 d1ff7f5
Correct depiction of Caliban logo (#45)
