### 2020-07-23 google/caliban#51 @arokem
`caliban cloud`: providing project ID through CLI fails
I am seeing:

```
$caliban cloud --nogpu mnist.py -- --project_id landscape-238422

No project_id found. 'caliban cloud' requires that you either set a
$PROJECT_ID environment variable with the ID of your Cloud project, or pass one
explicitly via --project_id. Try again, please!

```

Setting the environment variable does seem to work.
