### 2020-07-23 #2403 @arokem
I am trying to run `caliban cloud` and ran into https://github.com/google/caliban/issues/51.

Setting the environment variable seems to resolve that, but I then run into:

```
$caliban cloud --nogpu mnist.py
I0723 08:10:11.261835 4394732864 core.py:497] Generating Docker image with parameters:
I0723 08:10:11.262293 4394732864 core.py:498] {'adc_path': None,
 'build_path': '/Users/arokem/tmp/calibam/demo',
 'caliban_config': {},
 'conda_env_path': None,
 'credentials_path': None,
 'extra_dirs': None,
 'job_mode': <JobMode.CPU: 'CPU'>,
 'no_cache': False,
 'package': Package(executable=['python'], package_path='.', script_path='mnist.py', main_module=None),
 'requirements_path': 'requirements.txt',
 'setup_extras': None}
I0723 08:10:11.262627 4394732864 docker.py:614] Running command: docker build --rm -f- /Users/arokem/tmp/calibam/demo
Sending build context to Docker daemon  9.775kB
Step 1/11 : FROM gcr.io/blueshift-playground/blueshift:cpu
 ---> fafdb20241ad
Step 2/11 : RUN [ $(getent group 20) ] || groupadd --gid 20 20
 ---> Using cache
 ---> 0fb794a9d15b
Step 3/11 : RUN useradd --no-log-init --no-create-home -u 501 -g 20 --shell /bin/bash arokem
 ---> Using cache
 ---> 18bce73381b9
Step 4/11 : RUN mkdir -m 777 /usr/app /.creds /home/arokem
 ---> Using cache
 ---> 18c96e44e68c
Step 5/11 : ENV HOME=/home/arokem
 ---> Using cache
 ---> 7a4bb135f05e
Step 6/11 : WORKDIR /usr/app
 ---> Using cache
 ---> 266bb6485852
Step 7/11 : USER 501:20
 ---> Using cache
 ---> e3954e929d29
Step 8/11 : COPY --chown=501:20 requirements.txt /usr/app
 ---> Using cache
 ---> 08d9467b6ea9
Step 9/11 : RUN /bin/bash -c "pip install --no-cache-dir -r requirements.txt"
 ---> Using cache
 ---> d9eaf5a76590
Step 10/11 : COPY --chown=501:20 . /usr/app/.
 ---> Using cache
 ---> fdd5a32ee33f
Step 11/11 : ENTRYPOINT ["python", "mnist.py"]
 ---> Using cache
 ---> bbbce372b9f5
Successfully built bbbce372b9f5
The push refers to repository [gcr.io/landscape-238422/bbbce372b9f5]
45fcaa2b93ad: Preparing
3c1176c86461: Preparing
50e8c557ccbf: Preparing
bd07837419cb: Preparing
42d270e25e53: Preparing
c83c21629dcf: Waiting
a8c372c103ab: Waiting
4e84efa3db52: Waiting
25f9bffed627: Waiting
d4b51844f98b: Waiting
ed1d1d4a83ac: Waiting
df7bab7a7925: Waiting
28ba7458d04b: Waiting
838a37a24627: Waiting
a6ebef4a95c3: Waiting
b7f7d2967507: Waiting
denied: Token exchange failed for project 'landscape-238422'. Caller does not have permission 'storage.buckets.create'. To configure permissions, follow instructions at: https://cloud.google.com/container-registry/docs/access-control
Traceback (most recent call last):
  File "/Users/arokem/.virtualenvs/caliban/bin/caliban", line 8, in <module>
    sys.exit(main())
  File "/Users/arokem/.virtualenvs/caliban/lib/python3.7/site-packages/caliban/main.py", line 155, in main
    app.run(run_app, flags_parser=cli.parse_flags)
  File "/Users/arokem/.virtualenvs/caliban/lib/python3.7/site-packages/absl/app.py", line 299, in run
    _run_main(main, args)
  File "/Users/arokem/.virtualenvs/caliban/lib/python3.7/site-packages/absl/app.py", line 250, in _run_main
    sys.exit(main(argv))
  File "/Users/arokem/.virtualenvs/caliban/lib/python3.7/site-packages/caliban/main.py", line 145, in run_app
    xgroup=xgroup,
  File "/Users/arokem/.virtualenvs/caliban/lib/python3.7/site-packages/caliban/cloud/core.py", line 637, in submit_ml_job
    image_tag = generate_image_tag(project_id, docker_args, dry_run=dry_run)
  File "/Users/arokem/.virtualenvs/caliban/lib/python3.7/site-packages/caliban/cloud/core.py", line 505, in generate_image_tag
    image_tag = d.push_uuid_tag(project_id, image_id)
  File "/Users/arokem/.virtualenvs/caliban/lib/python3.7/site-packages/caliban/docker.py", line 655, in push_uuid_tag
    subprocess.run(["docker", "push", image_tag], check=True)
  File "/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/subprocess.py", line 468, in run
    output=stdout, stderr=stderr)
subprocess.CalledProcessError: Command '['docker', 'push', 'gcr.io/landscape-238422/bbbce372b9f5:latest']' returned non-zero exit status 1.
```

Do the permissions on the project need to be adjusted somehow?
