### 2020-07-23 google/caliban#51 @sritchie
Ah, yes, @arokem that is a subtlety with the argument parsing. Anything that comes after `--` is assumed to be an arg you want to pass directly into the container, to your script.

If you did this instead:

```
caliban cloud --nogpu --project_id landscape-238422 mnist.py
```

you should see it work out.

I'll be releasing an update today that removes the need for a service account key, which should make it slightly easier to get started.

let me know if this helps! Glad you got unblocked with the env variable.
