### 2020-07-23 google/caliban#49 @sritchie
This is really interesting. I didn't know about `REES`.

We actually DO now support `environment.yaml`; if you add one to your repo you'll get that environment installed. That needs a documentation update.

I didn't know about the `Dockerfile` ability. I wonder what that would do. This is a great tip. I'll look into this and keep notes here on the issue.

Thanks, @arokem !
