### 2020-07-23 google/caliban#48 @sritchie
👍 , I'll beef up `CONTRIBUTING.md` and make it clearer. Thank you!
