### 2020-07-23 d7e35d5
small improvements to documentation
