+++ Makefile
@@ -37,7 +37,7 @@ clean-files:
 .PHONY: install
 install:
 	rm -rf $(ENV_NAME)
-	virtualenv -p python3.6 $(ENV_NAME)
+	virtualenv -p python3 $(ENV_NAME)
 	$(PIP) install -r requirements-dev.txt
 	$(PIP) install -r docs/requirements.txt
 	$(PIP) install -e .
