### 2020-07-24 ca08a43
separate config, add julia version
