### 2020-07-24 f124926
split out project config
