### 2020-07-24 0bd5667
Add structure to project and global calibanconfig schema (#56)
