### 2020-07-29 a254611
add multiple mappings to tmpcopy
