### 2020-07-29 d414e58
fix case of tilde paths
