+++ caliban/util/fs.py
@@ -107,19 +107,34 @@ def generate_package(path: str,
 
 
 class TempCopy(object):
-  """Inside its scope, this class:
+  """TempCopy is a class that you can use as a context manager to transfer files
+  into the local directory with either
 
-  - generates a temporary file at tmp_name containing a copy of the file at
-    original_path, and
-  - deletes the new file at tmp_name when the scope exits.
+  - a name specified by you, or
+  - an autogenerated name.
 
-  The temporary file will live inside the current directory where python's
-  being executed; it's a hidden file, but it will be live for the duration of
-  TempCopy's scope.
+  Supply a dictionary mapping your files => your requested filename (or None).
+  When used as a context manager, TempCopy will
+
+  - copy data from every source to every destination
+  - return a dictionary of the source you provided => the destination's path in
+    the local directory.
+
+  The temporary files will live inside the current directory where python's
+  being executed for the duration of TempCopy's scope.
 
   We did NOT use a tmp directory here because the changing UUID name
   invalidates the docker image each time a new temp path / directory is
   generated.
+
+  mapping: dict mapping from optional src to optional destination filename. The
+           source should be a relative or absolute path on the local machine.
+           The destination should be a current-directory-local path, NOT an
+           absolute path. If a key is None, nothing will happen.
+
+  Returns:
+  dict of each non-None key in mapping => the destination's relative path in
+  the local directory.
 
   """
 
@@ -142,15 +157,12 @@ class TempCopy(object):
   def _sanitize_pair(self, src: Optional[str],
                      dst: Optional[str]) -> List[Tuple[str, str]]:
     """ If the source isn't none, returns a singleton with a pair of (from
-    directory, to_directory). Else, returns []."""
+    path, to_path). Else, returns []."""
 
     if src is None:
       return []
 
-    src = os.path.abspath(os.path.expanduser(src))
-    dst = dst or str(uuid.uuid1())
-
-    return [(src, dst)]
+    return [(src, dst or str(uuid.uuid1()))]
 
   def _sanitize_entries(
       self, mapping: Dict[Optional[str], Optional[str]]) -> Dict[str, str]:
@@ -166,9 +178,20 @@ class TempCopy(object):
     processed = [self._sanitize_pair(src, dst) for src, dst in mapping.items()]
     return dict(chain.from_iterable(processed))
 
-  def _prepend_dir(self, current_dir: str,
-                   mapping: Dict[str, str]) -> Dict[str, str]:
-    return {k: os.path.join(current_dir, v) for k, v in mapping.items()}
+  def _expand(self, current_dir: str, mapping: Dict[str,
+                                                    str]) -> Dict[str, str]:
+    """Expands the mapping that the user sees out into a mapping from absolute
+    source to absolute destination.
+
+    """
+
+    def final_src(src):
+      return os.path.abspath(os.path.expanduser(src))
+
+    def final_dst(dst):
+      return os.path.join(current_dir, dst)
+
+    return {final_src(k): final_dst(v) for k, v in mapping.items()}
 
   def __enter__(self) -> Dict[str, str]:
     """Entrypoint for the context manager. Performs the side effect of writing out
@@ -179,7 +202,8 @@ class TempCopy(object):
     """
     current_dir = os.getcwd()
     mapping = {**self._mapping}
-    to_write = self._prepend_dir(current_dir, mapping)
+
+    to_write = self._expand(current_dir, mapping)
 
     for src, dst in to_write.items():
       shutil.copy2(src, dst)
