### 2020-07-29 d5ddeae
Modify TempCopy to accept multiple mappings (#58)
