### 2020-07-29 5c9efd5
Add a clearer COPY command for better reuse in `caliban.docker.build` (#62)
