+++ caliban/docker/build.py
@@ -104,6 +104,28 @@ def apt_command(commands: List[str]) -> List[str]:
   return update + commands + cleanup
 
 
+def copy_command(user_id: int,
+                 user_group: int,
+                 from_path: str,
+                 to_path: str,
+                 comment: Optional[str] = None) -> str:
+  """Generates a Dockerfile entry that will copy the file at the directory-local
+  from_path into the container at to_path.
+
+  If you supply a relative path, Docker will copy the file into the current
+  working directory, where it will be overwritten in any interactive mode. We
+  recommend using an absolute path!
+
+  """
+  cmd = f"COPY --chown={user_id}:{user_group} {from_path} {to_path}"
+
+  if comment is not None:
+    comment_s = "\n# ".join(comment.split("\n"))
+    return f"# {comment_s}\n{cmd}"
+
+  return cmd
+
+
 # Dict linking a particular supported shell to the data required to run and
 # install the shell inside a container.
 #
@@ -224,15 +246,18 @@ def _dependency_entries(workdir: str,
   """
   ret = ""
 
+  def copy(from_path, to_path):
+    return copy_command(user_id, user_group, from_path, to_path)
+
   if setup_extras is not None:
     ret += f"""
-COPY --chown={user_id}:{user_group} setup.py {workdir}
+{copy("setup.py", workdir)}
 RUN /bin/bash -c "pip install --no-cache-dir {extras_string(setup_extras)}"
 """
 
   if conda_env_path is not None:
     ret += f"""
-COPY --chown={user_id}:{user_group} {conda_env_path} {workdir}
+{copy(conda_env_path, workdir)}
 RUN /bin/bash -c "{CONDA_BIN} env update \
     --quiet --name caliban \
     --file {conda_env_path} && \
@@ -241,7 +266,7 @@ RUN /bin/bash -c "{CONDA_BIN} env update \
 
   if requirements_path is not None:
     ret += f"""
-COPY --chown={user_id}:{user_group} {requirements_path} {workdir}
+{copy(requirements_path, workdir)}
 RUN /bin/bash -c "pip install --no-cache-dir -r {requirements_path}"
 """
 
@@ -259,26 +284,24 @@ def _package_entries(workdir: str, user_id: int, user_group: int,
   between files inside a project.
 
   """
-  owner = "{}:{}".format(user_id, user_group)
-
   arg = package.main_module or package.script_path
+  package_path = package.package_path
 
   # This needs to use json so that quotes print as double quotes, not single
   # quotes.
   entrypoint_s = json.dumps(package.executable + [arg])
-
-  return """
-# Copy project code into the docker container.
-COPY --chown={owner} {package_path} {workdir}/{package_path}
+  copy_code = copy_command(
+      user_id,
+      user_group,
+      package_path,
+      f"{workdir}/{package_path}",
+      comment="Copy project code into the docker container.")
+  return f"""
+{copy_code}
 
 # Declare an entrypoint that actually runs the container.
 ENTRYPOINT {entrypoint_s}
-  """.format_map({
-      "owner": owner,
-      "package_path": package.package_path,
-      "workdir": workdir,
-      "entrypoint_s": entrypoint_s
-  })
+  """
 
 
 def _service_account_entry(user_id: int, user_group: int, credentials_path: str,
@@ -298,26 +321,21 @@ def _service_account_entry(user_id: int, user_group: int, credentials_path: str,
   service account is present.
 
   """
-  container_creds = "{}/credentials.json".format(docker_credentials_dir)
-  ret = """
-COPY --chown={user_id}:{user_group} {credentials_path} {container_creds}
+  container_creds = f"{docker_credentials_dir}/credentials.json"
+  ret = f"""
[…]
