### 2020-07-29 af9dd99
Add support for running an embedded cloud_sql_proxy (#60)
