+++ caliban/resources/caliban_launcher.py
@@ -1,0 +1,139 @@
+#!/usr/bin/env python
+#
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+'''caliban wrapper utility'''
+
+import argparse
+import copy
+import json
+import logging
+import os
+import subprocess
+import sys
+import time
+
+logging.basicConfig(level=logging.INFO)
+
+
+def _parse_json(argname, json_string, expected_type):
+  """parses a json string, validating the return type"""
+
+  try:
+    obj = json.loads(json_string)
+    assert isinstance(obj, expected_type)
+  except Exception:
+    raise argparse.ArgumentTypeError(
+        "%s must be a json %s. Got '%s'." %
+        (argname, expected_type.__name__, json_string))
+
+  return obj
+
+
+def _parser():
+  '''generates argument parser'''
+
+  parser = argparse.ArgumentParser(
+      description='caliban wrapper for container.',
+      prog='caliban_wrapper',
+      formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+
+  parser.add_argument(
+      '--caliban_env',
+      type=lambda x: _parse_json('env', x, dict),
+      default={},
+      help='JSON dict of variables to set for container',
+  )
+
+  parser.add_argument(
+      '--caliban_command',
+      type=lambda x: _parse_json('command', x, list),
+      help='main container command to execute, as json list',
+  )
+
+  parser.add_argument(
+      '--caliban_service',
+      type=lambda x: _parse_json('service', x, list),
+      action='append',
+      help=
+      'run the given command as a service prior to running the main entrypoint')
+
+  parser.add_argument(
+      '--caliban_delay',
+      type=int,
+      default=5,
+      help='delay in seconds between service executions',
+  )
+
+  return parser
+
+
+# ----------------------------------------------------------------------------
+def _parse_flags(argv):
+  return _parser().parse_known_args(argv[1:])
+
+
+# ----------------------------------------------------------------------------
+def _start_services(services, env, delay):
+  '''runs the commands in the services list, returns a list of Popen instances
+  sets the environment variables in <env>, and delays by <delay> between
+  commands
+  '''
+  procs = []
+  for cmd in services:
+    logging.info(' '.join(cmd))
+    procs.append(subprocess.Popen(cmd, env=env))
+    time.sleep(delay)
+
+  return procs
+
+
+# ----------------------------------------------------------------------------
+def _execute_command(cmd, args, env):
+  '''executes the given command with the provided args and env vars
+  this blocks until the given command completes
+  '''
+
+  cmd = cmd + args
+  logging.info(' '.join(cmd))
+  subprocess.check_call(cmd, env=env)
+
+
+# ----------------------------------------------------------------------------
+def main(args, passthrough_args):
+
+  env = copy.copy(dict(os.environ))
+  caliban_env = args.caliban_env
+  cmd = args.caliban_command
+  services = args.caliban_service
+  delay = args.caliban_delay
+
+  logging.info('base env: %s' % str(env))
+  logging.info('env vars: %s' % str(caliban_env))
+  logging.info('command: %s' % ' '.join([str(x) for x in cmd]))
[…]
