### 2020-07-30 66da797
Update Julia install statements in Dockerfile
