### 2020-07-30 8b77859
Use julia_version instead of julia_url configuration key
