+++ caliban/docker/build.py
@@ -307,7 +307,7 @@ ENV JULIA_PROJECT={workdir}
 COPY --chown={user_id}:{user_group} *.toml {workdir}/
 COPY --chown={user_id}:{user_group} src {workdir}/src/
 RUN /bin/bash -c "cd {workdir} && \
-                  julia --eval 'using Pkg; Pkg.instantiate()' && \
+                  julia --eval 'using Pkg; Pkg.instantiate(); Pkg.precompile()' && \
                   rm -rf *"
 """
 
