### 2020-08-02 1b5e274
bugfix for https://github.com/google/caliban/issues/65: do not add resource maxima when quota is < 1
