### 2020-08-02 2d9826c
Add missing newline in Dockerfile
