### 2020-08-03 2bbeeac
minor fixes for missing CRLF, added traceback for execptions in cloud.core
