### 2020-08-03 3fd4d4f
fix for issue 68: adds newline in COPY directives generated by the copy_command method (#71)
This addresses issue #68, adding a newline to the end of COPY dockerfile directives generated by the copy_command() method in build.py. Thanks to @eschnett for finding this and the appropriate fix.
