### 2020-08-04 fc4b8e5
local runs working, creates non-job-specific config file for wrapper, wrapper using json args
