+++ caliban/util/metrics.py
@@ -25,11 +25,14 @@ import os
 
 import caliban.util as u
 import caliban.history.types as ht
-from typing import Dict, Any, List
+from contextlib import contextmanager
+from typing import Dict, Any, List, Optional
 
 CLOUD_SQL_WRAPPER_SCRIPT = 'cloud_sql_proxy.py'
-WRAPPER_SCRIPT = 'caliban_wrapper.py'
+WRAPPER_SCRIPT = 'caliban_launcher.py'
 RESOURCE_DIR = "/.resources"
+WRAPPER_CONFIG_FILE = 'caliban_wrapper_cfg.json'
+WRAPPER_CONFIG_PATH = os.path.join(RESOURCE_DIR, WRAPPER_CONFIG_FILE)
 
 
 def cloud_sql_proxy_path() -> str:
@@ -48,6 +51,81 @@ def wrapper_script_path() -> str:
   return u.resource(WRAPPER_SCRIPT)
 
 
+@contextmanager
+def wrapper_config_file(
+    path: str,
+    caliban_config: Optional[Dict[str, Any]] = None,
+):
+  '''creates a configuration file for the caliban wrapper script
+  This file contains the wrapper configuration that does not vary across
+  each caliban job being submitted, so it can be copied into the container.
+
+  This is to be used as a contextmanager yielding the path to the file:
+
+  with wrapper_config_file('.', caliban_config) as cfg_file:
+    # do things
+
+  The config file is deleted upon exiting the context scope.
+
+  Args:
+  path: directory in which to write file (this must exist
+  caliban_config: caliban configuration dictionary
+
+  Yields:
+  path to configuration file
+  '''
+
+  cfg = {}
+  mlflow_cfg = caliban_config.get('mlflow_config', None)
+  config_file_path = os.path.join(path, WRAPPER_CONFIG_FILE)
+
+  if mlflow_cfg is not None:
+    user = mlflow_cfg['user']
+    pw = mlflow_cfg['password']
+    db = mlflow_cfg['db']
+    project = mlflow_cfg['project']
+    region = mlflow_cfg['region']
+    artifact_root = mlflow_cfg['artifact_root']
+
+    socket_path = '/tmp/cloudsql'
+    proxy_path = os.path.join(os.sep, 'usr', 'bin', 'cloud_sql_proxy')
+
+    proxy_config = json.dumps({
+        'proxy': proxy_path,
+        'path': socket_path,
+        'project': project,
+        'region': region,
+        'db': db,
+        'creds': '~/.config/gcloud/application_default_credentials.json',
+    })
+
+    proxy_cmd = [
+        'python',
+        os.path.join(RESOURCE_DIR, CLOUD_SQL_WRAPPER_SCRIPT), proxy_config
+    ]
+
+    tracking_uri = (
+        f'postgresql+pg8000://{user}:{pw}@/{db}?unix_sock={socket_path}/'
+        f'{project}:{region}:{db}/.s.PGSQL.5432')
+
+    config = {
+        'services': [proxy_cmd],
+        'env': {
+            'MLFLOW_TRACKING_URI': tracking_uri,
+            'MLFLOW_ARTIFACT_ROOT': artifact_root
+        }
+    }
+
+  with open(config_file_path, 'w') as f:
+    json.dump(config, f, indent=2)
+
+  try:
+    yield config_file_path
+  finally:
+    if os.path.exists(config_file_path):
+      os.remove(config_file_path)
+
+
 def _mlflow_job_name(index: int, user: str = None) -> str:
   '''returns mlflow job name for local caliban job'''
   user = user or u.current_user()
[…]
