+++ tutorials/uv-metrics/requirements.txt
@@ -1,2 +1,0 @@
-uv-metrics
-pg8000==1.16.1
