+++ tutorials/uv-metrics/trainer/train.py
@@ -420,17 +420,19 @@ def model_main(activation='relu', width=1000, depth=2, lr=0.5, **kwargs):
   model = build_model(activation, width, depth)
   optimizer = tf.optimizers.SGD(lr)
 
-  with uv.start_run():
-    uv.report_params({
-        **kwargs,
-        **{
-            "depth": depth,
-            "width": width,
-            "lr": lr,
-            "activation": activation
-        }
-    })
-    train_and_log(model, optimizer, **kwargs)
+  uv.start_run()
+
+  MLFlowReporter().report_params({
+      **kwargs,
+      **{
+          "depth": depth,
+          "width": width,
+          "lr": lr,
+          "activation": activation
+      }
+  })
+
+  train_and_log(model, optimizer, **kwargs)
 
 
 def run_app(args):
