### 2020-08-04 202420f
use uv-metrics from master, remove requirements.txt and use only setup.py
