### 2020-08-05 e5398ad
clean up some unneeded diffs
