+++ setup.cfg
@@ -23,36 +23,3 @@ style = pep440
 versionfile_source = caliban/_version.py
 versionfile_build = caliban/_version.py
 tag_prefix =
-
-[mypy-tqdm.*]
-ignore_missing_imports = True
-
-[mypy-absl.*]
-ignore_missing_imports = True
-
-[mypy-blessings.*]
-ignore_missing_imports = True
-
-[mypy-googleapiclient.*]
-ignore_missing_imports = True
-
-[mypy-sqlalchemy.*]
-ignore_missing_imports = True
-
-[mypy-kubernetes.*]
-ignore_missing_imports = True
-
-[mypy-google.*]
-ignore_missing_imports = True
-
-[mypy-urllib3.*]
-ignore_missing_imports = True
-
-[mypy-yaspin.*]
-ignore_missing_imports = True
-
-[mypy-schema.*]
-ignore_missing_imports = True
-
-[mypy-commentjson.*]
-ignore_missing_imports = True
