### 2020-08-06 9d341fa
updated tutorial config to use new uv release, removed requirements.txt, minor change to train.py
