+++ tutorials/uv-metrics/setup.py
@@ -25,7 +25,7 @@ REQUIRED_PACKAGES = [
     "sqlalchemy==1.3.13",
     'tensorflow-cpu',
     'tensorflow_datasets',
-    'uv-metrics',
+    'uv-metrics>=0.4.2',
 ]
 
 setup(
