### 2020-08-06 1066156
updates to launcher script and util.metrics, added unit tests for new util.metrics methods
