+++ caliban/resources/caliban_launcher.py
@@ -13,7 +13,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-'''caliban wrapper utility'''
+'''This utility wraps the caliban user's entrypoint in order to configure
+additional service processes and environment variables for each caliban
+job.
+'''
 
 import argparse
 import copy
@@ -23,6 +26,10 @@ import os
 import subprocess
 import sys
 import time
+
+RESOURCE_DIR = "/.resources"
+LAUNCHER_CONFIG_FILE = 'caliban_launcher_cfg.json'
+LAUNCHER_CONFIG_PATH = os.path.join(RESOURCE_DIR, LAUNCHER_CONFIG_FILE)
 
 logging.basicConfig(level=logging.INFO)
 
@@ -42,18 +49,16 @@ def _parse_json(argname, json_string, expected_type):
 
 
 def _parser():
-  '''generates argument parser'''
-
   parser = argparse.ArgumentParser(
-      description='caliban wrapper for container.',
-      prog='caliban_wrapper',
+      description='caliban launcher for container.',
+      prog='caliban_launcher',
       formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 
   parser.add_argument(
-      '--caliban_env',
+      '--caliban_config',
       type=lambda x: _parse_json('env', x, dict),
       default={},
-      help='JSON dict of variables to set for container',
+      help='JSON config dictionary',
   )
 
   parser.add_argument(
@@ -62,30 +67,14 @@ def _parser():
       help='main container command to execute, as json list',
   )
 
-  parser.add_argument(
-      '--caliban_service',
-      type=lambda x: _parse_json('service', x, list),
-      action='append',
-      help=
-      'run the given command as a service prior to running the main entrypoint')
-
-  parser.add_argument(
-      '--caliban_delay',
-      type=int,
-      default=5,
-      help='delay in seconds between service executions',
-  )
-
   return parser
 
 
-# ----------------------------------------------------------------------------
 def _parse_flags(argv):
   return _parser().parse_known_args(argv[1:])
 
 
-# ----------------------------------------------------------------------------
-def _start_services(services, env, delay):
+def _start_services(services, env, delay=5):
   '''runs the commands in the services list, returns a list of Popen instances
   sets the environment variables in <env>, and delays by <delay> between
   commands
@@ -99,7 +88,6 @@ def _start_services(services, env, delay):
   return procs
 
 
-# ----------------------------------------------------------------------------
 def _execute_command(cmd, args, env):
   '''executes the given command with the provided args and env vars
   this blocks until the given command completes
@@ -110,14 +98,48 @@ def _execute_command(cmd, args, env):
   subprocess.check_call(cmd, env=env)
 
 
-# ----------------------------------------------------------------------------
+def _load_config_file():
+  '''loads the launcher configuration data from the config file
+  at ./resources/'caliban_laucher_cfg.json as a dict
+  '''
+  if not os.path.exists(LAUNCHER_CONFIG_PATH):
+    return {}
+
+  with open(LAUNCHER_CONFIG_PATH) as f:
+    cfg = json.load(f)
+
+  return cfg
+
+
+def _get_config(args):
+  '''gets the configuration dictionary for the launcher by combining the
+  static configuration in the launcher config file and the dynamic
+  configuration passed in the command args. Here the dynamic args take
+  precedence over static args where there is a collision.
+  '''
+
+  cfg = _load_config_file()
+  dynamic_config = args.caliban_config
+
[…]
