+++ caliban/util/metrics.py
@@ -29,10 +29,10 @@ from contextlib import contextmanager
 from typing import Dict, Any, List, Optional
 
 CLOUD_SQL_WRAPPER_SCRIPT = 'cloud_sql_proxy.py'
-WRAPPER_SCRIPT = 'caliban_launcher.py'
+LAUNCHER_SCRIPT = 'caliban_launcher.py'
 RESOURCE_DIR = "/.resources"
-WRAPPER_CONFIG_FILE = 'caliban_wrapper_cfg.json'
-WRAPPER_CONFIG_PATH = os.path.join(RESOURCE_DIR, WRAPPER_CONFIG_FILE)
+LAUNCHER_CONFIG_FILE = 'caliban_launcher_cfg.json'
+LAUNCHER_CONFIG_PATH = os.path.join(RESOURCE_DIR, LAUNCHER_CONFIG_FILE)
 GPU_ENABLED_TAG = 'gpu_enabled'
 TPU_ENABLED_TAG = 'tpu_enabled'
 JOB_NAME_TAG = 'job_name'
@@ -48,78 +48,102 @@ def cloud_sql_proxy_path() -> str:
   return u.resource(CLOUD_SQL_WRAPPER_SCRIPT)
 
 
-def wrapper_script_path() -> str:
-  """Returns an absolute path to the caliban_wrapper python script that we
+def launcher_path() -> str:
+  """Returns an absolute path to the caliban_launcher python script that we
   inject into containers.
 
   """
-  return u.resource(WRAPPER_SCRIPT)
+  return u.resource(LAUNCHER_SCRIPT)
+
+
+def _default_launcher_config() -> Dict[str, Any]:
+  return {
+      'services': [],
+      'env': {},
+  }
+
+
+def _create_mlflow_config(
+    cfg: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
+  '''generates mlflow configuration dict for launcher script
+  Args:
+  cfg: mlflow configuration dict from .calibanconfig.json
+
+  Returns:
+  config dict for launcher script with entries needed for mlflow
+  '''
+
+  if cfg is None:
+    return _default_launcher_config()
+
+  user = cfg['user']
+  pw = cfg['password']
+  db = cfg['db']
+  project = cfg['project']
+  region = cfg['region']
+  artifact_root = cfg['artifact_root']
+
+  socket_path = '/tmp/cloudsql'
+  proxy_path = os.path.join(os.sep, 'usr', 'bin', 'cloud_sql_proxy')
+
+  proxy_config = json.dumps({
+      'proxy': proxy_path,
+      'path': socket_path,
+      'project': project,
+      'region': region,
+      'db': db,
+      'creds': '~/.config/gcloud/application_default_credentials.json',
+  })
+
+  proxy_cmd = [
+      'python',
+      os.path.join(RESOURCE_DIR, CLOUD_SQL_WRAPPER_SCRIPT), proxy_config
+  ]
+
+  tracking_uri = (
+      f'postgresql+pg8000://{user}:{pw}@/{db}?unix_sock={socket_path}/'
+      f'{project}:{region}:{db}/.s.PGSQL.5432')
+
+  return {
+      'services': [proxy_cmd],
+      'env': {
+          'MLFLOW_TRACKING_URI': tracking_uri,
+          'MLFLOW_ARTIFACT_ROOT': artifact_root
+      }
+  }
 
 
 @contextmanager
-def wrapper_config_file(
+def launcher_config_file(
     path: str,
     caliban_config: Optional[Dict[str, Any]] = None,
 ):
-  '''creates a configuration file for the caliban wrapper script
-  This file contains the wrapper configuration that does not vary across
+  '''creates a configuration file for the caliban launcher script
[…]
