### 2020-08-06 5377b00
Add caliban launcher script to enable mlflow [in progress] (#63)
This PR:

Adds the ability to inject a launcher script into the container, with the ability to intercept arguments, set environment variables and start up a service that will be available for the duration of the container's task.
This is the conclusion of a port of #35 .

Co-authored-by: Ambrose Slone <aslone@google.com>
