### 2020-08-07 08b98cf
added debug flag for mlflow to reduce logging spam but still allow debugging connectivity issues
