### 2020-08-07 85fda3e
removing logging commands from launcher and sql proxy scripts
