### 2020-08-07 49fe8ae
Native MLFlow tracking server support (#35)
This PR adds support for MLFlow integration for local, google cloud-ai-platform, and google kubernetes engine jobs launched in Caliban. This adds a thin python launcher script to Caliban-generated docker image that performs tasks such as launching service processes and setting environment variables for container jobs.

Co-authored-by: Sam Ritchie <samritchie@google.com>
