+++ caliban/docker/build.py
@@ -20,11 +20,14 @@ and notebooks in a Docker environment.
 
 from __future__ import absolute_import, division, print_function
 
+import datetime
 import json
 import os
+import re
 import subprocess
 from enum import Enum
 from pathlib import Path
+from pkg_resources import resource_filename
 from typing import Any, Dict, List, NamedTuple, NewType, Optional, Union
 
 from absl import logging
@@ -33,6 +36,7 @@ from blessings import Terminal
 import caliban.config as c
 import caliban.util as u
 import caliban.util.fs as ufs
+import caliban.history.types as ht
 import caliban.util.metrics as um
 
 t = Terminal()
@@ -275,8 +279,71 @@ RUN /bin/bash -c "pip install --no-cache-dir -r {requirements_path}"
   return ret
 
 
-def _package_entries(workdir: str, user_id: int, user_group: int,
-                     package: u.Package) -> str:
+def _cloud_sql_proxy_entry(user_id: int, user_group: int) -> str:
+  '''returns dockerfile entry to fetch cloud_sql_proxy, installing in /usr/bin'''
+
+  cmd = (
+      'RUN wget -q https://dl.google.com/cloudsql/cloud_sql_proxy.linux.amd64 '
+      '-O /usr/bin/cloud_sql_proxy && '
+      'chmod 755 /usr/bin/cloud_sql_proxy')
+
+  return f'''
+USER root
+
+{cmd}
+
+USER {user_id}:{user_group}
+'''
+
+
+def _generate_entrypoint(
+    executable: str,
+    user_id: int,
+    user_group: int,
+    caliban_config: Optional[Dict[str, Any]] = None,
+) -> str:
+  '''generates dockerfile entry to set the container entrypoint, along with
+    any additional necessary entries, such as for services like cloud_sql_proxy
+
+  Args:
+  executable: string of main executable
+  caliban_config: dictionary of caliban configuration options
+  user_id: id of non-root user
+  user_group: id of non-root group for user
+
+  Returns:
+  string with Dockerfile directives to set ENTRYPOINT
+  '''
+
+  caliban_config = caliban_config or {}
+
+  mlflow_cfg = caliban_config.get('mlflow_config')
+  if mlflow_cfg is not None:
+    cloud_sql_proxy_code = _cloud_sql_proxy_entry(user_id=user_id,
+                                                  user_group=user_group)
+  else:
+    cloud_sql_proxy_code = ''
+
+  launcher_cmd = json.dumps([
+      'python',
+      os.path.join(RESOURCE_DIR, um.LAUNCHER_SCRIPT), '--caliban_command',
+      executable
+  ])
+
+  return f'''
+{cloud_sql_proxy_code}
+
+ENTRYPOINT {launcher_cmd}
+'''
+
+
+def _package_entries(
+    workdir: str,
+    user_id: int,
+    user_group: int,
+    package: u.Package,
+    caliban_config: Optional[Dict[str, Any]] = None,
+) -> str:
   """Returns the Dockerfile entries required to:
 
   - copy a directory of code into a docker container
@@ -286,24 +353,34 @@ def _package_entries(workdir: str, user_id: int, user_group: int,
   between files inside a project.
 
   """
+  caliban_config = caliban_config or {}
+
   arg = package.main_module or package.script_path
   package_path = package.package_path
 
-  # This needs to use json so that quotes print as double quotes, not single
-  # quotes.
-  entrypoint_s = json.dumps(package.executable + [arg])
   copy_code = copy_command(
       user_id,
       user_group,
       package_path,
       f"{workdir}/{package_path}",
       comment="Copy project code into the docker container.")
-  return f"""
-{copy_code}
 
[…]
