### 2020-08-08 ef7f1f4
added new A100 gpu type
