### 2020-08-09 8868afe
update accelerator table
