### 2020-08-09 868a2d9
update accelerator table (#75)
