### 2020-08-10 6b15f61
add contributor info (#76)
