### 2020-08-11 9247955
added support for different ubuntu base images to cloudbuild, added ubuntu 20.04 cpu images to build configuration (#78)
