### 2020-08-11 e7b5f1a
change default docker base image tags to reflect cloudbuild schema
