### 2020-08-12 aa3725d
fix to prevent mlflow env vars from being generated when there is no mlflow config
