+++ caliban/util/metrics.py
@@ -166,6 +166,7 @@ def _mlflow_job_name(index: int, user: str = None) -> str:
 
 
 def mlflow_args(
+    caliban_config: Dict[str, Any],
     experiment_name: str,
     index: int,
     tags: Dict[str, Any],
@@ -173,13 +174,17 @@ def mlflow_args(
   '''returns mlflow args for caliban launcher
   Args:
 
+  caliban_config: caliban configuration dict
   experiment: experiment object
-  index: job index
+  index: job index, if < 0, then no run name is generated
   tags: dictionary of tags to pass to mlflow
 
   Returns:
   mlflow args list
   '''
+
+  if caliban_config.get('mlflow_config') is None:
+    return []
 
   env = {f'ENVVAR_{k}': v for k, v in tags.items()}
   env['MLFLOW_EXPERIMENT_NAME'] = experiment_name
