### 2020-08-13 869f538
added additional tests for gke.util, fixed some minor bugs in gke.util
