+++ tests/caliban/platform/gke/test_util.py
@@ -16,11 +16,21 @@
 """unit tests for gke utilities"""
 import random
 import unittest
-from typing import List
+from typing import List, Optional, Dict
 from unittest import mock
+import uuid
+import json
+import os
+import tempfile
+import yaml
 
 import hypothesis.strategies as st
 from hypothesis import given, settings
+from kubernetes.client.api_client import ApiClient
+from kubernetes.client import V1Job
+import pytest
+import argparse
+import google
 
 import caliban.platform.cloud.types as ct
 import caliban.platform.gke.constants as k
@@ -72,6 +82,42 @@ class UtilTestSuite(unittest.TestCase):
       self.assertTrue(valid == (spec.count <= gpu_limits[spec.gpu]))
 
     return
+
+  # --------------------------------------------------------------------------
+  def test_validate_gpu_spec_against_limits_deterministic(self):
+    '''deterministic test to make sure we get full coverage'''
+
+    # gpu not supported
+    cfg = {
+        'gpu_spec': ct.GPUSpec(ct.GPU.K80, 1),
+        'gpu_limits': {
+            ct.GPU.P100: 1
+        },
+        'limit_type': 'zone',
+    }
+    assert not util.validate_gpu_spec_against_limits(**cfg)
+
+    # request above limit
+    cfg = {
+        'gpu_spec': ct.GPUSpec(ct.GPU.K80, 2),
+        'gpu_limits': {
+            ct.GPU.P100: 1,
+            ct.GPU.K80: 1,
+        },
+        'limit_type': 'zone',
+    }
+    assert not util.validate_gpu_spec_against_limits(**cfg)
+
+    # valid request
+    cfg = {
+        'gpu_spec': ct.GPUSpec(ct.GPU.K80, 1),
+        'gpu_limits': {
+            ct.GPU.P100: 1,
+            ct.GPU.K80: 1,
+        },
+        'limit_type': 'zone',
+    }
+    assert util.validate_gpu_spec_against_limits(**cfg)
 
   # --------------------------------------------------------------------------
   def test_nvidia_daemonset_url(self):
@@ -226,7 +272,7 @@ class UtilTestSuite(unittest.TestCase):
                                                'name',
                                                list(conds),
                                                0,
-                                               spinner=False))
+                                               spinner=True))
     else:
       self.assertIsNone(
           util.wait_for_operation(api, 'name', list(conds), 0, spinner=False))
@@ -311,6 +357,10 @@ class UtilTestSuite(unittest.TestCase):
 
     # idempotency check
     self.assertEqual(sanitized, util.sanitize_job_name(sanitized))
+
+    # ensure coverage, first char must be alnum, last must be alnum
+    x = '_' + sanitized + '-'
+    assert valid(util.sanitize_job_name(x))
 
     return
 
@@ -567,12 +617,16 @@ class UtilTestSuite(unittest.TestCase):
       values=everything(),
   ))
   def test_nonnull_dict(self, input_dict):
+    input_dict[str(uuid.uuid1())] = {'x': None, 'y': 7}  # ensure coverage
+    input_dict[str(uuid.uuid1())] = [1, 2, None, 3]
     self._validate_nonnull_dict(util.nonnull_dict(input_dict), input_dict)
     return
 
   # --------------------------------------------------------------------------
   @given(st.lists(everything()))
   def test_nonnull_list(self, input_list):
+    input_list.append({'x': None, 'y': 7})  # ensure coverage
[…]
