### 2020-08-13 05b6645
fix test for application_default_credentials_path
