### 2020-08-13 97f09e3
fix test for application_default_credentials_path
