### 2020-08-13 040e9b5
minor bug fixes for gke (#85)
This PR implements some minor bugfixes for the gke backend.
