+++ caliban/platform/gke/cluster.py
@@ -28,7 +28,7 @@ import urllib3
 import yaml
 from google.auth.credentials import Credentials
 from google.cloud.container_v1 import ClusterManagerClient
-from google.cloud.container_v1.types import NodePool
+from google.cloud.container_v1.types import NodePool, Cluster as GKECluster
 from googleapiclient import discovery
 from googleapiclient.http import HttpRequest
 from kubernetes.client import (V1Container, V1DaemonSet, V1EnvVar, V1Job,
@@ -188,13 +188,13 @@ class Cluster(object):
   """
 
   # --------------------------------------------------------------------------
-  def __init__(self, name: str, project_id: str, zone: str,
+  def __init__(self, name: Optional[str], project_id: str, zone: str,
                credentials: Credentials):
     self._cluster_client = None
-    self._gke_cluster = None
-    self._core_api = None
-    self._batch_api = None
-    self._apps_api = None
+    self._gke_cluster: Optional[GKECluster] = None
+    self._core_api: Optional[kubernetes.client.CoreV1Api] = None
+    self._batch_api: Optional[kubernetes.client.BatchV1Api] = None
+    self._apps_api: Optional[kubernetes.client.AppsV1Api] = None
     self._tpu_api = None
     self.name = name
     self.project_id = project_id
@@ -217,6 +217,9 @@ class Cluster(object):
     # if gke cluster info already populated, then noop
     # otherwise uses cluster api to get cluster info
     if not self._set_gke_cluster():
+      return False
+
+    if self._gke_cluster is None:
       return False
 
     # resolve our zone in case the wildcard '-' was passed
@@ -312,7 +315,7 @@ class Cluster(object):
 
     if client is None:
       logging.error('error getting cluster management client')
-      return False
+      return None
 
     clusters = util.get_gke_clusters(client, project_id, zone)
     return [c.name for c in clusters] if clusters is not None else None
@@ -347,7 +350,7 @@ class Cluster(object):
   def container_limits(
       accelerator: Optional[Accelerator],
       count: int = 1,
-      preemptible_tpu: bool = True) -> Optional[Dict[str, str]]:
+      preemptible_tpu: bool = True) -> Optional[Dict[str, int]]:
     """creates container limits dictionary for given accelerator type and count
 
     Args:
@@ -449,7 +452,7 @@ class Cluster(object):
       selector[k.NODE_SELECTOR_INSTANCE_TYPE] = machine_type.value
 
     # see: https://cloud.google.com/kubernetes-engine/docs/how-to/gpus
-    if type(accelerator) == GPU:
+    if isinstance(accelerator, GPU):
       selector[
           k.NODE_SELECTOR_GKE_ACCELERATOR] = accelerator.value.lower().replace(
               '_', '-')
@@ -493,6 +496,9 @@ class Cluster(object):
     list of V1Pod instances on success, None otherwise
     """
 
+    if self._core_api is None:
+      return None
+
     # this returns a V1PodList
     rsp = self._core_api.list_pod_for_all_namespaces(watch=False)
 
@@ -511,6 +517,9 @@ class Cluster(object):
     Returns:
     list of V1Job instances on success, None otherwise
     """
+    if self._batch_api is None:
+      return None
+
     return self._batch_api.list_job_for_all_namespaces(watch=False).items
 
   # --------------------------------------------------------------------------
@@ -518,6 +527,9 @@ class Cluster(object):
   @connected(None)
   def get_job(self, job_name: str) -> Optional[V1Job]:
     '''gets a v1job from the cluster from the given job name'''
+    if self._batch_api is None:
+      return None
+
     jobs = self._batch_api.list_job_for_all_namespaces(
         watch=False, field_selector=f'metadata.name={job_name}').items
     if len(jobs) != 1:
[…]
