### 2020-08-18 316d1a9
Slurm: Specify tasks and threads
