+++ caliban/docker/build.py
@@ -20,14 +20,11 @@ and notebooks in a Docker environment.
 
 from __future__ import absolute_import, division, print_function
 
-import datetime
 import json
 import os
-import re
 import subprocess
 from enum import Enum
 from pathlib import Path
-from pkg_resources import resource_filename
 from typing import Any, Dict, List, NamedTuple, NewType, Optional, Union
 
 from absl import logging
@@ -36,7 +33,6 @@ from blessings import Terminal
 import caliban.config as c
 import caliban.util as u
 import caliban.util.fs as ufs
-import caliban.history.types as ht
 import caliban.util.metrics as um
 
 t = Terminal()
@@ -281,62 +277,60 @@ RUN /bin/bash -c "pip install --no-cache-dir -r {requirements_path}"
   return ret
 
 
-def _cloud_sql_proxy_entry(user_id: int, user_group: int) -> str:
-  '''returns dockerfile entry to fetch cloud_sql_proxy, installing in /usr/bin'''
-
-  cmd = (
-      'RUN wget -q https://dl.google.com/cloudsql/cloud_sql_proxy.linux.amd64 '
-      '-O /usr/bin/cloud_sql_proxy && '
-      'chmod 755 /usr/bin/cloud_sql_proxy')
-
-  return f'''
-USER root
-
-{cmd}
-
-USER {user_id}:{user_group}
-'''
-
-
-def _generate_entrypoint(
-    executable: str,
+def _cloud_sql_proxy_entry(
     user_id: int,
     user_group: int,
     caliban_config: Optional[Dict[str, Any]] = None,
 ) -> str:
-  '''generates dockerfile entry to set the container entrypoint, along with
-    any additional necessary entries, such as for services like cloud_sql_proxy
+  """returns dockerfile entry to fetch cloud_sql_proxy, installing in /usr/bin.
+
+  Args:
+  user_id: id of non-root user
+  user_group: id of non-root group for user
+  caliban_config: dictionary of caliban configuration options
+
+  Returns:
+  string with Dockerfile directives to install cloud_sql_proxy as root
+  and reset the user to the specified user_id:user_group.
+
+  """
+  caliban_config = caliban_config or {}
+  mlflow_cfg = caliban_config.get('mlflow_config')
+
+  if mlflow_cfg is None:
+    return ""
+
+  return f"""
+USER root
+
+RUN wget \
+  -q https://dl.google.com/cloudsql/cloud_sql_proxy.linux.amd64 \
+  -O /usr/bin/cloud_sql_proxy \
+  && chmod 755 /usr/bin/cloud_sql_proxy
+
+USER {user_id}:{user_group}
+"""
+
+
+def _generate_entrypoint(executable: str) -> str:
+  """generates dockerfile entry to set the container entrypoint.
 
   Args:
   executable: string of main executable
-  caliban_config: dictionary of caliban configuration options
-  user_id: id of non-root user
-  user_group: id of non-root group for user
 
   Returns:
   string with Dockerfile directives to set ENTRYPOINT
-  '''
 
-  caliban_config = caliban_config or {}
-
-  mlflow_cfg = caliban_config.get('mlflow_config')
-  if mlflow_cfg is not None:
-    cloud_sql_proxy_code = _cloud_sql_proxy_entry(user_id=user_id,
-                                                  user_group=user_group)
-  else:
-    cloud_sql_proxy_code = ''
-
+  """
   launcher_cmd = json.dumps([
       'python',
       os.path.join(RESOURCE_DIR, um.LAUNCHER_SCRIPT), '--caliban_command',
       executable
   ])
 
-  return f'''
-{cloud_sql_proxy_code}
-
+  return f"""
 ENTRYPOINT {launcher_cmd}
-'''
+"""
 
 
 def _package_entries(
[…]
