+++ caliban/docker/build.py
@@ -22,6 +22,7 @@ from __future__ import absolute_import, division, print_function
 
 import json
 import os
+import re
 import subprocess
 from enum import Enum
 from pathlib import Path
