### 2020-09-03 #2403 @lukasheinrich
Hi @sritchie

I successfully ran through the examples and went through the checklist and am generally satisfied (all boxes are checked)

Generally I think it would be useful if the demo tutorial was included in the docs (and not link back out to github) and agree with @arokem that it would be useful to cite binder/repo2docker.

Apologies again for the delay. This seems like a very useful piece of software.
