### 2020-09-03 #2403 @sritchie
Thanks @lukasheinrich! I did mention boot2docker and binder in related work, so thanks for that:

![image](https://user-images.githubusercontent.com/69635/92117245-2e00fe00-edb2-11ea-910b-f184da7353db.png)

Yes, I think you're probably right about mirroring the tutorial. I bet I can do some readthedocs magic and get the docs site to import the README wholesale...
