### 2020-09-04 53c05c3
added non-none project check to caliban launcher
