### 2020-09-04 #2403 @diehlpk
> I successfully ran through the examples and went through the checklist and am generally satisfied (all boxes are checked)

@lukasheinrich It seems that

`Functionality: Have the functional claims of the software been confirmed?`

is not checked. Could you please verify?
