### 2020-09-04 #2403 @diehlpk
> Yes, I think you're probably right about mirroring the tutorial. I bet I can do some readthedocs magic and get the docs site to import the README wholesale...

@sritchie Let me know once you accomplished that.
