### 2020-09-08 31784c7
added tests for caliban_launcher.py
