### 2020-09-10 6428aa5
modified .calibanconfig.json schema for uv/mlflow/pubsub
