+++ tests/caliban/util/test_metrics.py
@@ -171,8 +171,12 @@ def test_launcher_config_file():
                 'user': user,
                 'password': password,
                 'artifact_root': artifact_root,
-                'pubsub_topic': 'mlflow',
             },
+            'uv': {
+                'mlflow': {
+                    'pubsub_topic': 'mlflow',
+                }
+            }
         },
     }
 
