+++ caliban/config/__init__.py
@@ -137,8 +137,15 @@ MLFlowConfig = {
     'password': str,
     'artifact_root': str,
     s.Optional('debug'): bool,
-    s.Optional('pubsub_project'): str,  # default = project above
+}
+
+UVMLFlowConfig = {
+    s.Optional('pubsub_project'): str,  # default = mlflow_config.project
     s.Optional('pubsub_topic', default='mlflow'): str,
+}
+
+UVConfig = {
+    s.Optional('mlflow'): UVMLFlowConfig,
 }
 
 # Config items that are project-specific, and don't belong in a global
@@ -161,6 +168,7 @@ SystemConfig = {
     s.Optional("default_mode", default=JobMode.CPU): s.Use(JobMode.parse),
     s.Optional("gcloud", default={}): GCloudConfig,
     s.Optional("mlflow_config", default=None): MLFlowConfig,
+    s.Optional("uv", default={}): UVConfig,
 }
 
 # The final, parsed calibanconfig.
