### 2020-09-12 #2403 @sritchie
- [x]  Make a tagged release of your software, and list the version tag of the archived version here: 0.4.1, https://github.com/google/caliban/releases/tag/0.4.1
- [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository): [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4026612.svg)](https://doi.org/10.5281/zenodo.4026612)
- [x]  Please list the DOI of the archived version here: `10.5281/zenodo.4026612`, [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4026612.svg)](https://doi.org/10.5281/zenodo.4026612)

@diehlpk , I'm not going to take the suggestion at the moment to mirror the tutorial from the README to the readthedocs page, as it's slightly more involved, and I need to think through a proper publishing workflow to guarantee that users who find the README can get started right away without the information falling out of sync.

Let me know if I can do anything else here!
