### 2020-09-16 #2403 @kyleniemeyer
Hi @sritchie, I'm doing some final checks on your submission. In the paper, it looks like the Forde et al. and Zaharia et al. references may be missing some details. Also, please look at the other references, and add curly braces `{}` around words to enforce capitalization where appropriate (e.g., "IEEE", "Linux").
