+++ paper/paper.bib
@@ -1,16 +1,23 @@
 @article{merkel2014docker,
-  title={Docker: lightweight linux containers for consistent development and deployment},
-  author={Merkel, Dirk},
-  journal={Linux journal},
-  volume={2014},
-  number={239},
-  pages={2},
-  year={2014}
+author = {Merkel, Dirk},
+title = {Docker: {L}ightweight {L}inux {C}ontainers for {C}onsistent {D}evelopment and {D}eployment},
+year = {2014},
+issue_date = {March 2014},
+publisher = {Belltown Media},
+address = {Houston, TX},
+volume = {2014},
+number = {239},
+issn = {1075-3583},
+abstract = {Docker promises the ability to package applications and their dependencies into lightweight containers that move easily between different distros, start up quickly and are isolated from each other.},
+journal = {Linux J.},
+month = mar,
+articleno = {2},
+numpages = {1}
 }
 
 @inproceedings{cito2016,
 author = {Cito, J\"{u}rgen and Gall, Harald C.},
-title = {Using Docker Containers to Improve Reproducibility in Software Engineering Research},
+title = {Using {D}ocker {C}ontainers to {I}mprove {R}eproducibility in {S}oftware {E}ngineering {R}esearch},
 year = {2016},
 isbn = {9781450342056},
 publisher = {Association for Computing Machinery},
@@ -28,7 +35,7 @@ series = {ICSE ’16}
 @inproceedings{deng2009imagenet,
   title={Imagenet: A large-scale hierarchical image database},
   author={Deng, Jia and Dong, Wei and Socher, Richard and Li, Li-Jia and Li, Kai and Fei-Fei, Li},
-  booktitle={2009 IEEE conference on computer vision and pattern recognition},
+  booktitle={2009 {IEEE} conference on computer vision and pattern recognition},
   pages={248--255},
   url={https://doi.org/10.1109/cvpr.2009.5206848},
   doi={10.1109/cvpr.2009.5206848},
@@ -37,14 +44,17 @@ series = {ICSE ’16}
 }
 
 @article{zaharia2018accelerating,
-  title={Accelerating the Machine Learning Lifecycle with MLflow.},
-  author={Zaharia, Matei and Chen, Andrew and Davidson, Aaron and Ghodsi, Ali and Hong, Sue Ann and Konwinski, Andy and Murching, Siddharth and Nykodym, Tomas and Ogilvie, Paul and Parkhe, Mani and others},
-  year={2018}
+  title={Accelerating the {M}achine {L}earning {L}ifecycle with {MLflow}},
+  author={M. Zaharia and Andrew Chen and A. Davidson and A. Ghodsi and S. Hong and A. Konwinski and Siddharth Murching and Tomas Nykodym and P. Ogilvie and Mani Parkhe and F. Xie and Corey Zumar},
+  journal={{IEEE} Data Eng. Bull.},
+  year={2018},
+  volume={41},
+  pages={39-45}
 }
 
-@inproceedings{Forde2018ReproducingML,
-  title={Reproducing Machine Learning Research on Binder},
-  author={Jessica Forde and Matthias Bussonnier and F{\'e}lix-Antoine Fortin and Brian E. Granger and Tim Head and Chris Holdgraf and Paul Ivanov and Kyle Kelley and Michael Pacer and Yuvi Panda and Fernando P{\'e}rez and Gladys Nalvarte and Benjamin Ragan-Kelley and Zachary R. Sailer and Steven Silvester and Erik Sundell and Carol Willing},
+@article{Forde2018ReproducingML,
+  title={Reproducing {M}achine {L}earning {R}esearch on {B}inder},
[…]
