### 2020-09-16 #2403 @diehlpk
@sritchie For the second reference ieee should be IEEE.
