+++ paper/paper.bib
@@ -40,7 +40,7 @@ series = {ICSE ’16}
   url={https://doi.org/10.1109/cvpr.2009.5206848},
   doi={10.1109/cvpr.2009.5206848},
   year={2009},
-  organization={Ieee}
+  organization={IEEE}
 }
 
 @article{zaharia2018accelerating,
@@ -52,9 +52,10 @@ series = {ICSE ’16}
   pages={39-45}
 }
 
-@article{Forde2018ReproducingML,
+@inproceedings{Forde2018ReproducingML,
   title={Reproducing {M}achine {L}earning {R}esearch on {B}inder},
   author={Forde, Jessica and Bussonnier, Matthias and Fortin, F{\'e}lix-Antoine and Granger, Brian and Head, Tim and Holdgraf, Chris and Ivanov, Paul and Kelley, Kyle and Pacer, M and Panda, Yuvi and others},
+  booktitle={{NIPS} {W}orkshop on {M}achine {L}earning {O}pen {S}ource {S}oftware},
   year={2018}
 }
 
