### 2020-09-16 #2403 @kyleniemeyer
Hey @sritchie, looks like the Forde et al. reference is still missing something. Where was that published—a journal, or on a website, or something?
