### 2020-09-16 #2403 @sritchie
@kyleniemeyer , I took the bibtex entry from here: https://scholar.google.com/scholar?q=Reproducing+%7BMachine+%7BLearning+%7BResearch+on+%7BBinder&hl=en&as_sdt=0&as_vis=1&oi=scholart:

```
@article{forde2018reproducing,
  title={Reproducing Machine Learning Research on Binder},
  author={Forde, Jessica and Bussonnier, Matthias and Fortin, F{\'e}lix-Antoine and Granger, Brian and Head, Tim and Holdgraf, Chris and Ivanov, Paul and Kelley, Kyle and Pacer, M and Panda, Yuvi and others},
  year={2018}
}
```

![image](https://user-images.githubusercontent.com/69635/93403272-40644880-f844-11ea-9726-bc6581a7be6a.png)

It lives on OpenReview.net: https://openreview.net/forum?id=BJlR6KTE3X

and was published at NIPS 2018 Workshop MLOSS; you're right that even though that isn't in the bibtex entry Scholar generates it needs to be here. Adding now.
