### 2020-09-17 #2403 @kyleniemeyer
@sritchie yeah, unfortunately Google Scholar's BiBTeX generation only seems to work consistently for regular journal articles.
